% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_compute_pip_stats.R
\name{ag_compute_pip_stats}
\alias{ag_compute_pip_stats}
\title{Computes poverty statistics (aggregated)}
\usage{
ag_compute_pip_stats(
  welfare,
  povline,
  population,
  area,
  area_pop,
  requested_mean,
  popshare = NULL,
  default_ppp = NULL,
  ppp = NULL,
  p0 = 0.5
)
}
\arguments{
\item{welfare}{numeric: Cumulative proportion of welfare held by that
proportion of the population (Lorenz Curve).}

\item{povline}{numeric: Poverty line.}

\item{population}{numeric: Cumulative proportion of population.}

\item{area}{character: Area (Urban or Rural)}

\item{area_pop}{numeric: Total population per area.}

\item{requested_mean}{numeric: Welfare mean.}

\item{popshare}{numeric: Share of population living below the poverty line.
Optional.}

\item{default_ppp}{numeric: Default purchasing power parity.}

\item{ppp}{numeric: PPP request by user.}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}
}
\value{
list
}
\description{
Compute poverty statistics for aggregated data distribution.
}
\keyword{internal}
