% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{BETAI}
\alias{BETAI}
\title{BETAI}
\usage{
BETAI(a, b, x)
}
\arguments{
\item{a}{numeric}

\item{b}{numeric}

\item{x}{numeric}
}
\value{
numeric
}
\description{
BETAI
}
\examples{
BETAI(a, b, x)

}
