% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_dist_stats.R
\name{check_curve_validity_dist_lb}
\alias{check_curve_validity_dist_lb}
\title{check_curve_validity_dist_lb
Check validity of Lorenz beta fit for distributional statistics}
\usage{
check_curve_validity_dist_lb(A, B, C)
}
\arguments{
\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[3]}.}
}
\value{
list
}
\description{
check_curve_validity_dist_lb
Check validity of Lorenz beta fit for distributional statistics
}
\keyword{internal}
