% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_utils.R
\name{check_get_gd_fun_params}
\alias{check_get_gd_fun_params}
\title{Check parameters of get_gd functions}
\usage{
check_get_gd_fun_params(lp)
}
\arguments{
\item{lp}{list of parameters}
}
\value{
invisible TRUE
}
\description{
Check parameters of get_gd functions
}
