% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{check_input_gd_compute_pip_stats_lq}
\alias{check_input_gd_compute_pip_stats_lq}
\title{Title}
\usage{
check_input_gd_compute_pip_stats_lq(
  population,
  welfare,
  mean,
  default_ppp,
  povline,
  ppp,
  popshare,
  is_lq,
  p0
)
}
\arguments{
\item{population}{numeric: cumulative proportion of population.}

\item{welfare}{numeric: cumulative proportion of income held by that
proportion of the population (Lorenz Curve).}

\item{mean}{numeric: Welfare mean}

\item{default_ppp}{numeric: Default purchasing power parity}

\item{povline}{numeric: Poverty line}

\item{ppp}{numeric: PPP request by user}

\item{popshare}{numeric: Share of population living below the poverty line.
Optional}

\item{is_lq}{boolean: indicates whether the estimation of lorenz quadratic}
}
\value{

}
\description{
Title
}
