% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_neg_values}
\alias{check_neg_values}
\title{Check for negative values in vector}
\usage{
check_neg_values(x)
}
\value{
A message if any values are negative
}
\description{
It is used for now on Lorenz curves equation
}
\keyword{internal}
