% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_median.R
\name{compute_median.pipgd}
\alias{compute_median.pipgd}
\title{calculate the median for group data}
\usage{
\method{compute_median}{pipgd}(dt, welfare, weight, mean, p0 = 0.5, pop = NULL)
}
\arguments{
\item{dt}{dataframe with group data}

\item{welfare}{numeric: Cumulative proportion of welfare held by that
proportion of the population (Lorenz Curve).}

\item{weight}{population weights vector}

\item{mean}{numeric. Named vector of welfare mean for each data level
available in data frame \code{dt}}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}

\item{level}{character: data level to filter data. It assumes variable
\code{max_domain} in data set}
}
\value{
scalar
}
\description{
calculate the median for group data
}
