% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_median.R
\name{compute_median.pipmd}
\alias{compute_median.pipmd}
\title{compute median for microdata}
\usage{
\method{compute_median}{pipmd}(dt, welfare, weight, level = NULL)
}
\arguments{
\item{dt}{data frame with either micro data, group data, or imputed data}

\item{welfare}{welfare vector}

\item{weight}{population weights vector}

\item{level}{character: data level to filter data. It assumes variable
\code{max_domain} in data set}

\item{x}{dataframe with microdata}
}
\value{
scalar
}
\description{
compute median for microdata
}
