% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_median.R
\name{compute_median}
\alias{compute_median}
\title{compute median for different methods}
\usage{
compute_median(dt, ...)
}
\arguments{
\item{dt}{data frame with either micro data, group data, or imputed data}

\item{welfare}{welfare vector}

\item{weight}{population weights vector}
}
\value{

}
\description{
compute median for different methods
}
