% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_median.R
\name{compute_median_gd}
\alias{compute_median_gd}
\title{calculate the median for group data}
\usage{
compute_median_gd(dt, welfare, weight, mean, p0 = 0.5, level = NULL)
}
\arguments{
\item{welfare}{numeric: Cumulative proportion of welfare held by that
proportion of the population (Lorenz Curve).}

\item{weight}{population weights vector}

\item{mean}{numeric: Welfare mean.}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}

\item{c}{dataframe with group data}
}
\value{
scalar
}
\description{
calculate the median for group data
}
