% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_median.R
\name{compute_median_md}
\alias{compute_median_md}
\title{compute median for microdata}
\usage{
compute_median_md(dt, welfare, weight, level = NULL)
}
\arguments{
\item{welfare}{welfare vector}

\item{weight}{population weights vector}

\item{x}{dataframe with microdata}
}
\value{
scalar
}
\description{
compute median for microdata
}
