% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_request_year_mean.R
\name{compute_predicted_mean}
\alias{compute_predicted_mean}
\title{compute_predicted_mean}
\usage{
compute_predicted_mean(survey_mean, proxy)
}
\arguments{
\item{survey_mean}{numeric: A vector with one or two survey means.}

\item{proxy}{list: A list with proxy values.}
}
\value{
numeric
}
\description{
compute_predicted_mean
}
