% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_spl.R
\name{compute_spl.pipgd}
\alias{compute_spl.pipgd}
\title{Compute societal poverty lines for grouped data in PIP}
\usage{
\method{compute_spl}{pipgd}(
  dt,
  welfare,
  weight,
  mean,
  cpi = pipload::pip_load_aux("cpi"),
  ppp = pipload::pip_load_aux("ppp"),
  pop = NULL
)
}
\arguments{
\item{dt}{data frame with either micro data, group data, or imputed data}

\item{welfare}{character: name of variable with welfare vector}

\item{weight}{character: name of variable with population (weight) vector}

\item{cpi}{data frame with CPI data. Default \code{pipload::pip_load_aux("cpi")}}

\item{ppp}{data frame with PPP data. Default \code{pipload::pip_load_aux("ppp")}}
}
\description{
Compute societal poverty lines for grouped data in PIP
}
