% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_spl.R
\name{compute_spl.pipmd}
\alias{compute_spl.pipmd}
\title{Compute societal poverty lines for microdata in PIP}
\usage{
\method{compute_spl}{pipmd}(
  dt,
  welfare,
  weight,
  cpi = pipload::pip_load_aux("cpi"),
  ppp = pipload::pip_load_aux("ppp")
)
}
\arguments{
\item{dt}{data frame with either micro data, group data, or imputed data}

\item{welfare}{character: name of variable with welfare vector}

\item{weight}{character: name of variable with population (weight) vector}

\item{cpi}{data frame with CPI data. Default \code{pipload::pip_load_aux("cpi")}}

\item{ppp}{data frame with PPP data. Default \code{pipload::pip_load_aux("ppp")}}
}
\description{
Compute societal poverty lines for microdata in PIP
}
