% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_spl.R
\name{compute_spl}
\alias{compute_spl}
\title{Compute SPL}
\usage{
compute_spl(weighted_median_ppp, threshold_rate = 0.5, min_level = 1.9)
}
\arguments{
\item{weighted_median_ppp}{numeric: Median. Weighted median in PPP terms.}

\item{threshold_rate}{numeric: A value between 0 and 1 for the proportion of
the median welfare for SPL calculation.}

\item{min_level}{numeric: Minimum level of the SPL. Defaults to the current
IPL ($1.90).}
}
\value{
numeric
}
\description{
Compute the societal poverty line based on the median welfare in PPP-adjusted
dollars.
}
\references{
D. Jolliffe, E. B. Prydz. 2017. "\href{https://documents1.worldbank.org/curated/en/133671495562984832/pdf/Societal-poverty-a-relative-and-relevant-measure.pdf}{Societal Poverty: A Relative and Relevant Measure}".
Policy Research Working Paper 8073. World Bank, Washington, DC.
}
\keyword{internal}
