% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{create_functional_form_lb}
\alias{create_functional_form_lb}
\title{create_functional_form_lb}
\usage{
create_functional_form_lb(welfare, population)
}
\arguments{
\item{welfare}{numeric: Welfare vector from empirical Lorenz curve}

\item{population}{numeric: Population vector from empirical Lorenz curve}
}
\value{
data.frame
}
\description{
Prepare data for Lorenz beta regression.
}
\details{
Prepares data for regression. The last observation of (p,l), which by
construction has the value (1, 1), is excluded since the functional form for
the Lorenz curve already forces it to pass through the point (1, 1).
}
\references{
Kakwani, N. 1980. "\href{https://EconPapers.repec.org/RePEc:ecm:emetrp:v:48:y:1980:i:2:p:437-46}{On a Class of Poverty Measures}".
\emph{Econometrica 48} (2): 437-46.
}
\keyword{internal}
