% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{create_functional_form_lq}
\alias{create_functional_form_lq}
\title{Prepares data for Lorenz Quadratic regression}
\usage{
create_functional_form_lq(welfare, population)
}
\arguments{
\item{welfare}{numeric: Welfare vector from empirical Lorenz curve.}

\item{population}{numeric: Population vector from empirical Lorenz curve.}
}
\value{
list: a) a numeric vector 'y', b) data frame with three columns x1, x2, x3
}
\description{
Prepares data for regression of L(1-L) on (P^2-L), L(P-1) and (P-L). The last
observation of (p,l), which by construction has the value (1, 1), is excluded
since the functional form for the Lorenz curve already forces it to pass
through the point (1, 1). Equation 15 in Lorenz Quadratic original paper.
}
\references{
Krause, M. 2013. "\href{https://doi.org/10.1016/j.jeconom.2013.01.001}{Corrigendum to Elliptical Lorenz curves}". \emph{Journal of
Econometrics 174} (1): 44.

Villasenor, J., B. C. Arnold. 1989. "\href{https://EconPapers.repec.org/RePEc:eee:econom:v:40:y:1989:i:2:p:327-338}{Elliptical Lorenz curves}".
\emph{Journal of Econometrics 40} (2): 327-338.
}
\keyword{internal}
