% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{DDLK}
\alias{DDLK}
\title{DDLK}
\usage{
DDLK(h, A, B, C)
}
\arguments{
\item{h}{numeric}

\item{A}{numeric vector: lorenz curve coefficient}

\item{B}{numeric vector: lorenz curve coefficient}

\item{C}{numeric vector: lorenz curve coefficient}
}
\value{
numeric
}
\description{
DDLK
}
\examples{
DDLK(h, A, B, C)

}
