% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflate_welfare_mean.R
\name{deflate_welfare_mean}
\alias{deflate_welfare_mean}
\title{Deflate welfare mean}
\usage{
deflate_welfare_mean(welfare_mean, ppp, cpi)
}
\arguments{
\item{welfare_mean}{numeric: The weighted welfare mean from a survey.}

\item{ppp}{numeric: PPP value}

\item{cpi}{numeric: CPI value.}
}
\value{
numeric
}
\description{
Adjust the welfare mean of a survey to constant PPP-adjusted dollars to make
surveys comparable over time and across countries.
}
\details{
Note that the CPI value to be used must be pre-adjusted so that welfare
aggregates are comparable across surveys within a given country. This implies
that adjustments to account for temporal and spatial factors within the
survey as well as currency changes between surveys must already have been
made.
}
\references{
Azevedo, J. P., P. A. Corral, D. Jolliffe, C. Lakner, D. G. Mahler,
J. Montes, M. C. Nguyen, E. B. Prydz. 2018.
"\href{http://documents1.worldbank.org/curated/en/651541537208471889/pdf/129963-WP-PUBLIC-Disclosed-9-19-2018.pdf}{Prices used in Global Poverty Measurement}".
Global Poverty Monitoring Technical Note 3.
World Bank, Washington, DC.
}
