% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{expand_components}
\alias{expand_components}
\title{Expand component into parent function}
\usage{
expand_components(ct)
}
\arguments{
\item{ct}{list: components from \code{get_components_lq}.}
}
\value{

}
\description{
Expand component into parent function
}
