% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_gini.R
\name{fgini}
\alias{fgini}
\title{Not compiled Gini}
\usage{
fgini(welfare, weight)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights.}
}
\value{
scalar
}
\description{
Not compiled Gini
}
