% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgt_lq.R
\name{fgt_lq}
\alias{fgt_lq}
\title{Title}
\usage{
fgt_lq(A, B, C, mu, z, ct)
}
\arguments{
\item{mu}{numeric: distributional mean}

\item{z}{numeric: Poverty line}

\item{ct}{list of parametric quadratic Lorenz components (from \code{get_components_lq()})}
}
\value{

}
\description{
Title
}
