% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_gaps.R
\name{fill_gaps}
\alias{fill_gaps}
\title{Fill gaps}
\usage{
fill_gaps(
  request_year,
  data = list(df0, df1 = NULL),
  survey_mean,
  survey_year,
  proxy = list(value0, value1 = NULL, req_value),
  data_type = c("microdata", "groupdata"),
  poverty_line = 1.9
)
}
\arguments{
\item{request_year}{integer: A value with the request year.}

\item{data}{list: A list with one or two data frames containing survey data.}

\item{survey_mean}{numeric: A vector with one or two survey means.}

\item{survey_year}{numeric: A vector with one or two survey years.}

\item{proxy}{list: A list with proxy values to be used to estimate the growth
rate.}

\item{data_type}{character: Type of data, either micro or grouped.}

\item{poverty_line}{numeric: Daily poverty line in international dollars.}
}
\description{
Calculate poverty statistics for a common request year by applying the World
Bank official method for interpolating and extrapolating surveys.
}
\examples{

}
