% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_quantile_functions.R
\name{format_out}
\alias{format_out}
\title{Format for output on quantile functions}
\usage{
format_out(quantiles, format = "atomic", name = "values")
}
\arguments{
\item{format}{character: "dt", "list", "atomic", giving the format of
the output. Default: "atomic"}

\item{name}{Name of the column that contains the key values from output
if format is "dt"}

\item{quantiles:}{the output from functions \code{md_quantile_values},
\code{md_welfare_share_at}, and \code{md_quantile_welfare_share}.}
}
\value{
output of quantile functions: see \code{format}.
}
\description{
This function helps modify the output of quantile functions to a specific
format (see \code{format} parameter for the options)
}
\keyword{internal}
