% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_dist_stats.R
\name{gd_compute_dist_fit_lb}
\alias{gd_compute_dist_fit_lb}
\title{Computes the sum of squares of error
Measures the fit of the model to the data for distributional statistics.}
\usage{
gd_compute_dist_fit_lb(welfare, population, A, B, C)
}
\arguments{
\item{welfare}{numeric: Welfare vector (grouped).}

\item{population}{numeric: Population vector (grouped).}

\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[3]}.}
}
\value{
list
}
\description{
Computes the sum of squares of error
Measures the fit of the model to the data for distributional statistics.
}
\keyword{internal}
