% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_dist_stats.R
\name{gd_compute_dist_stats}
\alias{gd_compute_dist_stats}
\title{Computes distributional statistics (grouped)}
\usage{
gd_compute_dist_stats(welfare, population, mean, p0 = 0.5)
}
\arguments{
\item{welfare}{numeric: Cumulative proportion of welfare held by that
proportion of the population (Lorenz Curve).}

\item{population}{numeric: Cumulative proportion of population.}

\item{mean}{numeric: Welfare mean. (default 1)}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}
}
\value{
list
}
\description{
Compute distributional statistics for grouped data by selecting the best
functional fit for the Lorenz curve (either beta or quadratic).
}
\examples{
# Compute distributional statistics
res <- wbpip:::gd_compute_dist_stats(
 welfare = grouped_data_ex2$welfare,
 population = grouped_data_ex2$weight,
 mean = 50)

}
\keyword{internal}
