% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_dist_stats_lq}
\alias{gd_compute_dist_stats_lq}
\title{Computes distributional stats from Lorenz Quadratic fit}
\usage{
gd_compute_dist_stats_lq(
  mean,
  p0,
  A,
  B,
  C,
  ppp_year = getOption("wbpip.available_ppp_years")
)
}
\arguments{
\item{mean}{numeric: Welfare mean.}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}

\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[3]}.}

\item{ppp_year}{numeric PPP year - 2017 (default) or 2011}
}
\value{
list
}
\description{
Computes distributional stats from Lorenz Quadratic fit
}
\keyword{internal}
