% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_median.R
\name{gd_compute_median}
\alias{gd_compute_median}
\title{calculate the median for group data}
\usage{
gd_compute_median(welfare, population, mean, p0 = 0.5)
}
\arguments{
\item{welfare}{numeric: Cumulative proportion of welfare held by that
proportion of the population (Lorenz Curve).}

\item{population}{numeric: Cumulative proportion of population.}

\item{mean}{numeric: Welfare mean.}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}
}
\value{
scalar
}
\description{
calculate the median for group data
}
\keyword{internal}
