% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_mld_lq}
\alias{gd_compute_mld_lq}
\title{Computes MLD from Lorenz Quadratic fit}
\usage{
gd_compute_mld_lq(A, B, C, key_values)
}
\arguments{
\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[3]}.}

\item{key_values}{named list: key values for lq fit, calculated using A, B, C
parameters using \link{gd_lq_key_values}}
}
\value{
numeric
}
\description{
\code{gd_compute_mld_lq()} computes the Mean Log deviation (MLD) from a Lorenz
Quadratic fit
}
