% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats.R
\name{gd_compute_pip_stats}
\alias{gd_compute_pip_stats}
\title{Computes poverty and inequality statistics (grouped)}
\usage{
gd_compute_pip_stats(
  welfare,
  povline = 1,
  population,
  mean = 1,
  popshare = NULL,
  default_ppp = 1,
  ppp = NULL,
  p0 = 0.5
)
}
\arguments{
\item{welfare}{numeric: Cumulative proportion of welfare held by that
proportion of the population (Lorenz Curve).}

\item{povline}{numeric: Poverty line. (default 1)}

\item{population}{numeric: Cumulative proportion of population.}

\item{mean}{numeric: Welfare mean. (default 1)}

\item{popshare}{numeric: Share of population living below the poverty line.
Optional.}

\item{default_ppp}{numeric: Default purchasing power parity.}

\item{ppp}{numeric: PPP request by user.}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}
}
\value{
list
}
\description{
Compute poverty statistics for grouped data by selecting the best functional
fit for the Lorenz curve (either beta or quadratic).
}
\examples{
# Compute PIP stats
res <- wbpip::gd_compute_pip_stats(
         grouped_data_ex2$welfare,
         grouped_data_ex2$weight,
         mean = 2.911786,
         povline = 1.9,
         default_ppp = 1)

}
