% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{gd_compute_pip_stats_lb}
\alias{gd_compute_pip_stats_lb}
\title{Computes poverty statistics (Lorenz beta)}
\usage{
gd_compute_pip_stats_lb(
  welfare,
  povline,
  population,
  requested_mean,
  popshare = NULL,
  default_ppp,
  ppp = NULL,
  p0 = 0.5,
  ppp_year = getOption("wbpip.available_ppp_years")
)
}
\arguments{
\item{welfare}{numeric: Cumulative proportion of welfare held by that
proportion of the population (Lorenz Curve).}

\item{povline}{numeric: Poverty line.}

\item{population}{numeric: Cumulative proportion of population.}

\item{requested_mean}{numeric: Welfare mean.}

\item{popshare}{numeric: Share of population living below the poverty line.
Optional.}

\item{default_ppp}{numeric: Default purchasing power parity.}

\item{ppp}{numeric: PPP request by user.}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}

\item{ppp_year}{numeric PPP year - 2017 (default) or 2011}
}
\value{
list
}
\description{
Compute poverty statistics for grouped data using the beta functional form of
the Lorenz qurve.
}
\keyword{#}
\keyword{1)}
\keyword{1.9,}
\keyword{50,}
\keyword{<-}
\keyword{=}
\keyword{Beta}
\keyword{Estimate}
\keyword{Lorenz}
\keyword{default_ppp}
\keyword{grouped_data_ex2$weight,}
\keyword{grouped_data_ex2$welfare,}
\keyword{internal}
\keyword{lb}
\keyword{population}
\keyword{povline}
\keyword{requested_mean}
\keyword{wbpip:::gd_compute_pip_stats_lb(}
\keyword{welfare}
