% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gd_compute_polarization}
\alias{gd_compute_polarization}
\title{Computes polarization index from parametric Lorenz fit}
\usage{
gd_compute_polarization(mean, p0, dcm, A, B, C)
}
\arguments{
\item{mean}{numeric: Welfare mean}

\item{p0}{numeric: To document}

\item{dcm}{numeric: To document}

\item{A}{numeric: Lorenz curve coefficient}

\item{B}{numeric: Lorenz curve coefficient}

\item{C}{numeric: Lorenz curve coefficient}
}
\value{
numeric
}
\description{
Used for grouped data computations
}
