% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{gd_compute_pov_gap_lb}
\alias{gd_compute_pov_gap_lb}
\title{Compute poverty gap for Lorenz Beta fit}
\usage{
gd_compute_pov_gap_lb(mean, povline, headcount, A, B, C, u = NULL)
}
\arguments{
\item{headcount}{numeric: Headcount index.}

\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[3]}.}

\item{u}{numeric: Normalized mean.}
}
\value{
numeric
}
\description{
Compute poverty gap for Lorenz Beta fit
}
\keyword{internal}
