% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_poverty_stats.R
\name{gd_compute_poverty_stats}
\alias{gd_compute_poverty_stats}
\title{Computes poverty statistics (grouped)}
\usage{
gd_compute_poverty_stats(
  welfare,
  povline,
  population,
  requested_mean,
  popshare = NULL,
  default_ppp = NULL,
  ppp = NULL,
  p0 = 0.5
)
}
\arguments{
\item{welfare}{numeric: Cumulative proportion of welfare held by that
proportion of the population (Lorenz Curve).}

\item{povline}{numeric: Poverty line.}

\item{population}{numeric: Cumulative proportion of population.}

\item{requested_mean}{numeric: Welfare mean.}

\item{popshare}{numeric: Share of population living below the poverty line.
Optional.}

\item{default_ppp}{numeric: Default purchasing power parity.}

\item{ppp}{numeric: PPP request by user.}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}
}
\value{
list
}
\description{
Compute poverty statistics for grouped data by selecting the best functional
fit for the Lorenz curve (either beta or quadratic).
}
\examples{
# Compute poverty stats
res <- wbpip:::gd_compute_poverty_stats(welfare = pipfaker::datt$lwelfare,
                                        population = pipfaker::datt$weight,
                                        povline = 1.9 * 365 / 12,
                                        requested_mean = 109.9,
                                        default_ppp = 1)
}
\keyword{internal}
