% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{gd_compute_poverty_stats_lb}
\alias{gd_compute_poverty_stats_lb}
\title{Computes poverty stats from Lorenz beta fit}
\usage{
gd_compute_poverty_stats_lb(mean, povline, A, B, C)
}
\arguments{
\item{mean}{numeric: Welfare mean.}

\item{povline}{numeric: Poverty line.}

\item{A}{numeric: First regression coefficient.}

\item{B}{numeric: Second regression coefficient.}

\item{C}{numeric: Third regression coefficient.}
}
\value{
list
}
\description{
Computes poverty stats from Lorenz beta fit
}
\keyword{internal}
