% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_poverty_stats_lq}
\alias{gd_compute_poverty_stats_lq}
\title{Computes poverty stats from Lorenz Quadratic fit}
\usage{
gd_compute_poverty_stats_lq(mean, povline, A, B, C)
}
\arguments{
\item{mean}{numeric: Welfare mean.}

\item{povline}{numeric: Poverty line.}

\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[3]}.}

\item{s1}{numeric: \strong{TO BE DOCUMENTED}.}

\item{s2}{numeric: \strong{TO BE DOCUMENTED}.}
}
\value{
list
}
\description{
Computes poverty stats from Lorenz Quadratic fit
}
\keyword{internal}
