% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{gd_compute_prosperity_gap_lb}
\alias{gd_compute_prosperity_gap_lb}
\title{Prosperity Gap with Lorenz Beta}
\usage{
gd_compute_prosperity_gap_lb(
  welfare,
  population,
  mean,
  pop = NULL,
  p0 = 0.5,
  nobs = 1e+05,
  cons_floor = c(0.5),
  verbose = FALSE
)
}
\arguments{
\item{welfare}{numeric: Cumulative proportion of welfare held by that
proportion of the population (Lorenz Curve).}

\item{population}{numeric: Cumulative proportion of population.}

\item{mean}{numeric: Welfare mean.}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}

\item{nobs}{numeric: Number of observations to be used in synthetic vector}

\item{cons_floor}{numeric: gives the consumption floor. Default is $0.5 per day, if NULL then no floor.}
}
\description{
Use the beta parameterization of the Lorenz curve to calculate the prosperity gap on group data
}
