% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_spl_lq}
\alias{gd_compute_spl_lq}
\title{Compute societal poverty line for group data from Lorenz quadratic fit}
\usage{
gd_compute_spl_lq(
  A = NULL,
  B = NULL,
  C = NULL,
  median = NULL,
  ppp_year = getOption("wbpip.available_ppp_years")
)
}
\arguments{
\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[3]}.}

\item{median}{numeric median welfare}

\item{ppp_year}{numeric PPP year}
}
\description{
Compute societal poverty line for group data from Lorenz quadratic fit
}
