% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_dist_stats.R
\name{gd_estimate_dist_stats_lb}
\alias{gd_estimate_dist_stats_lb}
\title{Estimates distributional stats from beta Lorenz fit}
\usage{
gd_estimate_dist_stats_lb(mean, p0, A, B, C, ppp_year)
}
\arguments{
\item{mean}{numeric: Welfare mean.}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}

\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[3]}.}

\item{ppp_year}{numeric PPP year - 2017 (default) or 2011}
}
\value{
list
}
\description{
Estimates distributional stats from beta Lorenz fit
}
\keyword{internal}
