% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_estimate_distribution.R
\name{gd_estimate_distribution}
\alias{gd_estimate_distribution}
\title{Estimate distribution (grouped data)}
\usage{
gd_estimate_distribution(
  welfare,
  population,
  gd_type,
  mean,
  povline,
  popshare = NULL
)
}
\arguments{
\item{welfare}{numeric: Cumulative proportion of welfare held by that
proportion of the population (Lorenz Curve).}

\item{population}{numeric: Cumulative proportion of population.}

\item{gd_type}{integer: Type of data distribution
If \code{gd_type = 1} \code{population} must be the cumulative proportion of
population and \code{welfare} must be the cumulative proportion of income held by
that proportion of the population (Lorenz curve). If \code{gd_type = 2},
\code{population} must be the proportion of population and \code{welfare} must be the
proportion of income. If \code{gd_type = 5}, then \code{population} must be the
Percentage of the population in a given interval of incomes, whereas
\code{welfare} must be the mean income of that interval.}

\item{mean}{numeric: Data mean}

\item{povline}{numeric: Poverty line. (default 1)}

\item{popshare}{numeric: Share of population living below the poverty line.
Optional.}
}
\value{
data.frame
}
\description{
Estimate distribution (grouped data)
}
\examples{
welfare <- c(24.84, 35.8, 45.36, 55.1, 64.92, 77.08, 91.75, 110.64, 134.9,
167.76, 215.48, 261.66, 384.97)
population <- c(0.92, 2.47, 5.11, 7.9, 9.69, 15.24, 13.64, 16.99, 10, 9.78,
3.96, 1.81, 2.49)

gd_estimate_distribution(welfare = welfare,
population = population,
gd_type = 5,
mean = 109.9,
povline = 89)
}
