% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_poverty_stats.R
\name{gd_estimate_poverty_stats_lb}
\alias{gd_estimate_poverty_stats_lb}
\title{Estimates poverty stats from beta Lorenz fit}
\usage{
gd_estimate_poverty_stats_lb(mean, povline, A, B, C)
}
\arguments{
\item{mean}{numeric: Welfare mean.}

\item{povline}{numeric: Poverty line.}

\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[3]}.}
}
\value{
list
}
\description{
Estimates poverty stats from beta Lorenz fit
}
\keyword{internal}
