% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_poverty_stats.R
\name{gd_estimate_poverty_stats_lq}
\alias{gd_estimate_poverty_stats_lq}
\title{gd_estimate_poverty_stats_lq
Estimates poverty stats from Quadratic Lorenz fit}
\usage{
gd_estimate_poverty_stats_lq(mean, povline, A, B, C)
}
\arguments{
\item{mean}{numeric: Welfare mean.}

\item{povline}{numeric: Poverty line.}

\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[3]}.}
}
\value{
list
}
\description{
gd_estimate_poverty_stats_lq
Estimates poverty stats from Quadratic Lorenz fit
}
\keyword{internal}
