% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gd_GHI_2009_income}
\alias{gd_GHI_2009_income}
\title{GHI 2009 Income (group data)}
\format{
A data frame with 10 rows and 5 variables:
\describe{
\item{country_code}{A fake three letter country code.}
\item{survey_year}{The year of the survey.}
\item{weight}{The weight of each indvidual's welfare value.}
\item{welfare}{The welfare of each indvidual. Can represent either income or consumption.}
\item{area}{Geographical location; urban or rural.} }
}
\source{
Worldbank internal.
}
\usage{
gd_GHI_2009_income
}
\description{
A dataset containing national group data for the fake country GHI for the
year 2009. Income based welfare values.
}
\details{
Note that nothing in this dataset is real. It is only meant for testing and
example purposes.
}
\keyword{datasets}
