% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_dist_stats.R
\name{gd_lq_key_values}
\alias{gd_lq_key_values}
\title{Calculates the key values in Table 2 of Datt (1998)}
\usage{
gd_lq_key_values(A, B, C)
}
\arguments{
\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[3]}.}
}
\value{
list
}
\description{
Calculates the key values in Table 2 of Datt (1998)
}
\keyword{internal}
