% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_dist_stats.R
\name{gd_select_lorenz_dist}
\alias{gd_select_lorenz_dist}
\title{gd_select_lorenz_dist
Select best Lorenz fit for distributional statistics}
\usage{
gd_select_lorenz_dist(lq, lb)
}
\arguments{
\item{lq}{list: Results from Lorenz Quadratic functional form. output of
\code{gd_compute_pip_stats_lq()}.}

\item{lb}{list: Results from Lorenz Beta functional form. output of
\code{gd_compute_pip_stats_lb()}.}
}
\value{
list
}
\description{
gd_select_lorenz_dist
Select best Lorenz fit for distributional statistics
}
\keyword{internal}
