% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_select_lorenz.R
\name{gd_select_lorenz_pov}
\alias{gd_select_lorenz_pov}
\title{Select the Lorenz functional to compute poverty statistics}
\usage{
gd_select_lorenz_pov(lq, lb)
}
\arguments{
\item{lq}{list: Results from Lorenz Quadratic functional form. output of
\code{gd_compute_pip_stats_lq()}.}

\item{lb}{list: Results from Lorenz Beta functional form. output of
\code{gd_compute_pip_stats_lb()}.}
}
\value{
logical
}
\description{
Select the Lorenz functional to compute poverty statistics
}
