% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low-hanging_indicators.R
\name{get_average_shortfall}
\alias{get_average_shortfall}
\title{Calculate average shortfall for the poor using headcount, poverty gap ratio, and poverty line}
\usage{
get_average_shortfall(headcount, povgap, povline)
}
\arguments{
\item{headcount}{Numeric vector giving headcount ratio, i.e. P0 FGT measure}

\item{povgap}{Numeric vector giving poverty gap ratio, i.e. P1 FGT measure}

\item{povline}{Numeric vector giving the poverty line}
}
\value{
Numeric vector giving average shortfall for the poor in the units of the poverty line
}
\description{
Calculate average shortfall for the poor using headcount, poverty gap ratio, and poverty line
}
\examples{
if (suppressWarnings(require(pipr))) {
  pip <- get_stats()
  get_average_shortfall(
    pip$headcount,
    pip$poverty_gap,
    pip$poverty_line) |>
    head()
}
}
