% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{get_components_lq}
\alias{get_components_lq}
\title{Title}
\usage{
get_components_lq(A, B, C)
}
\arguments{
\item{A}{numeric: First regression coefficient}

\item{B}{numeric: Second regression coefficient}

\item{C}{numeric: Third regression coefficient}
}
\value{

}
\description{
Title
}
