% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gd.R
\name{get_gd_lorenz_params}
\alias{get_gd_lorenz_params}
\title{Get Lorenz Parameters}
\usage{
get_gd_lorenz_params(welfare, population)
}
\arguments{
\item{welfare}{numeric: cumulative sahre of welfare (income/consumption)}

\item{population}{numeric: cumulative share of the population}
}
\value{
list with parameters
}
\description{
Get Parameters and key values derived from the quadratic and Beta Lorenz
parametrization. \code{welfare} and \code{population} must be vectors of a group data
dataset
}
\examples{
# Get Lorenz parameters
res <- get_gd_lorenz_params(
  welfare = grouped_data_ex2$welfare,
  population = grouped_data_ex2$weight)
str(res)
}
