% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gd.R
\name{get_gd_lorenz_stats}
\alias{get_gd_lorenz_stats}
\title{Get vectors related to the Lorenz Curve}
\usage{
get_gd_lorenz_stats(
  welfare = NULL,
  population = NULL,
  params = NULL,
  mean = 1,
  times_mean = 1,
  povline = mean * times_mean,
  complete = FALSE
)
}
\arguments{
\item{welfare}{numeric: cumulative sahre of welfare (income/consumption)}

\item{population}{numeric: cumulative share of the population}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}
}
\value{
data.table or list with Lorenz stats
}
\description{
Get vectors related to the Lorenz Curve
}
