% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gd.R
\name{get_gd_lorenz_validity_dist}
\alias{get_gd_lorenz_validity_dist}
\title{Check valirity of Lorenz Curve for distriburional Stats}
\usage{
get_gd_lorenz_validity_dist(
  welfare = NULL,
  population = NULL,
  params = NULL,
  complete = FALSE
)
}
\arguments{
\item{welfare}{numeric: cumulative sahre of welfare (income/consumption)}

\item{population}{numeric: cumulative share of the population}

\item{params}{list of parameters from \code{get_gd_lorenz_params()}}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}
}
\value{
list of distributional validity of each Lorenz model
}
\description{
Check valirity of Lorenz Curve for distriburional Stats
}
