% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gd.R
\name{get_gd_pov_gap_nv}
\alias{get_gd_pov_gap_nv}
\title{Group Data Poverty Gap}
\usage{
get_gd_pov_gap_nv(
  welfare = NULL,
  population = NULL,
  params = NULL,
  mean = 1,
  times_mean = 1,
  povline = mean * times_mean,
  complete = FALSE,
  lorenz = NULL
)
}
\arguments{
\item{welfare}{numeric: cumulative sahre of welfare (income/consumption)}

\item{population}{numeric: cumulative share of the population}

\item{params}{list of parameters from \code{get_gd_lorenz_validity()}}

\item{mean}{numeric: welfare mean of distribution.}

\item{povline}{numeric: value of poverty line. Default is the \code{mean} value}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}

\item{lorenz}{character or NULL. Lorenz curve selected. It could be "lq" for
Lorenz Quadratic or "lb" for Lorenz Beta}
}
\value{
data.table, list or atomic vector
}
\description{
Group Data Poverty Gap
}
