% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gd.R
\name{get_gd_select_lorenz}
\alias{get_gd_select_lorenz}
\title{Get selected Lorenz curve for distributional stats}
\usage{
get_gd_select_lorenz(
  welfare = NULL,
  population = NULL,
  params = NULL,
  mean = 1,
  times_mean = 1,
  popshare = NULL,
  povline = ifelse(is.null(popshare), mean * times_mean, NULL),
  complete = FALSE
)
}
\arguments{
\item{welfare}{numeric: cumulative sahre of welfare (income/consumption)}

\item{population}{numeric: cumulative share of the population}

\item{params}{list of parameters from \code{get_gd_lorenz_validity()}}

\item{mean}{numeric: welfare mean of distribution.}

\item{popshare}{numeric: range (0,1). Share of population. Provide share of
population instead of poverty line}

\item{povline}{numeric: value of poverty line. Default is the \code{mean} value}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}
}
\value{
list of values with best lorenz fit for distributional Stats
}
\description{
Get selected Lorenz curve for distributional stats
}
\examples{
# Using Lorenz parameters from get_gd_lorenz_params
params <- get_gd_lorenz_params(
  welfare = grouped_data_ex2$welfare,
  population = grouped_data_ex2$weight)

params <- get_gd_lorenz_validity(
  params = params,
  complete = TRUE)
get_gd_select_lorenz(params = params)

# Using Lorenz parameters from get_gd_lorenz_validity
params <- get_gd_lorenz_validity(
  welfare = grouped_data_ex2$welfare,
  population = grouped_data_ex2$weight,
  complete = TRUE)
get_gd_select_lorenz(params = params)

# Using original vectors

get_gd_select_lorenz(
  welfare = grouped_data_ex2$welfare,
  population = grouped_data_ex2$weight)
}
