% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gd.R
\name{get_gd_selected_lorenz_dist}
\alias{get_gd_selected_lorenz_dist}
\title{Get selected Lorenz curve for distributional stats}
\usage{
get_gd_selected_lorenz_dist(
  welfare = NULL,
  population = NULL,
  params = NULL,
  complete = FALSE
)
}
\arguments{
\item{welfare}{numeric: cumulative sahre of welfare (income/consumption)}

\item{population}{numeric: cumulative share of the population}

\item{params}{list of parameters from \code{get_gd_lorenz_params()}}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}
}
\value{
list of values with best lorenz fit for distributional Stats
}
\description{
Get selected Lorenz curve for distributional stats
}
\examples{
# Using Lorenz parameters from get_gd_lorenz_params
params <- get_gd_lorenz_params(
  welfare = grouped_data_ex2$welfare,
  population = grouped_data_ex2$weight)

params <- get_gd_lorenz_validity_dist(
  params = params,
  complete = TRUE)
get_gd_selected_lorenz_dist(params = params)

# Using Lorenz parameters from get_gd_lorenz_validity_dist
params <- get_gd_lorenz_validity_dist(
  welfare = grouped_data_ex2$welfare,
  population = grouped_data_ex2$weight,
  complete = TRUE)
get_gd_selected_lorenz_dist(params = params)

# Using original vectors

get_gd_selected_lorenz_dist(
  welfare = grouped_data_ex2$welfare,
  population = grouped_data_ex2$weight)
}
