% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gd.R
\name{get_gd_selected_lorenz_pov}
\alias{get_gd_selected_lorenz_pov}
\title{Get selected Lorenz curve for distributional stats}
\usage{
get_gd_selected_lorenz_pov(
  welfare = NULL,
  population = NULL,
  params = NULL,
  mean = 1,
  povline = mean,
  complete = FALSE
)
}
\arguments{
\item{welfare}{numeric: cumulative sahre of welfare (income/consumption)}

\item{population}{numeric: cumulative share of the population}

\item{params}{list of parameters from \code{get_gd_lorenz_params()}}

\item{mean}{numeric: welfare mean of distribution.}

\item{povline}{numeric: value of poverty line. Default is the \code{mean} value}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}
}
\description{
Get selected Lorenz curve for distributional stats
}
