% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gd.R
\name{get_gd_wlf_sahre_by_qtl}
\alias{get_gd_wlf_sahre_by_qtl}
\title{Welfare share by quantile in group data}
\usage{
get_gd_wlf_sahre_by_qtl(
  welfare = NULL,
  population = NULL,
  params = NULL,
  complete = FALSE,
  lorenz = NULL,
  n = 10
)
}
\arguments{
\item{welfare}{numeric: cumulative sahre of welfare (income/consumption)}

\item{population}{numeric: cumulative share of the population}

\item{params}{list of parameters from \code{get_gd_lorenz_validity()}}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}

\item{lorenz}{character or NULL. Lorenz curve selected. It could be "lq" for
Lorenz Quadratic or "lb" for Lorenz Beta}

\item{n}{numeric: Number of quantiles}
}
\value{
list with vector of share of welfare by quantiles
}
\description{
Welfare share by quantile in group data
}
\examples{
# Using params
params <- get_gd_select_lorenz(
  welfare = grouped_data_ex2$welfare,
  population = grouped_data_ex2$weight)
qt <- get_gd_wlf_sahre_by_qtl(params = params)
qt$dist_stats$quantiles

# Using orignal vectors
qt <- get_gd_wlf_sahre_by_qtl(
welfare = grouped_data_ex2$welfare,
  population = grouped_data_ex2$weight)
qt$dist_stats$quantiles
}
