% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low-hanging_indicators.R
\name{get_lh_9010_ratio}
\alias{get_lh_9010_ratio}
\title{Calculate ratio of the top 10\% to the bottom 10\% of the welfare distribution}
\usage{
get_lh_9010_ratio(top10, bottom10)
}
\arguments{
\item{top10}{Numeric vector giving the welfare of the 10th decile - i.e. the top 10\%}

\item{bottom10}{Numeric vector giving the welfare of the 1st decile - i.e. the bottom 10\%}
}
\value{
Numeric vector giving the ratio of the top 10\% to the bottom 10\% of the welfare distribution
}
\description{
Calculate ratio of the top 10\% to the bottom 10\% of the welfare distribution
}
\examples{
if (suppressWarnings(require(pipr))) {
  pip <- get_stats()

  get_lh_9010_ratio(
    top10    = pip$decile10,
    bottom10 = pip$decile1
    ) |>
    head()
}
}
