% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low-hanging_indicators.R
\name{get_lh_income_gap_ratio}
\alias{get_lh_income_gap_ratio}
\title{Calculate income gap ratio from headcount and poverty gap}
\usage{
get_lh_income_gap_ratio(headcount, povgap)
}
\arguments{
\item{headcount}{Numeric vector giving headcount ratio, i.e. P0 FGT measure}

\item{povgap}{Numeric vector giving poverty gap ratio, i.e. P1 FGT measure}
}
\value{
Numeric vector giving the income gap ratio
}
\description{
Calculate income gap ratio from headcount and poverty gap
}
\examples{
if (suppressWarnings(require(pipr))) {
  pip <- get_stats()
  get_lh_income_gap_ratio(
    pip$headcount,
    pip$poverty_gap) |>
    head()
}
}
