% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low-hanging_indicators.R
\name{get_lh_palma_ratio}
\alias{get_lh_palma_ratio}
\title{Calculate Palma ratio using decile information}
\usage{
get_lh_palma_ratio(
  top10,
  bottom40 = NULL,
  decile1 = NULL,
  decile2 = NULL,
  decile3 = NULL,
  decile4 = NULL
)
}
\arguments{
\item{top10}{Numeric vector giving either welfare or welfare share for tenth decile}

\item{bottom40}{numeric: sum of deciles 1 to 4.}

\item{decile1}{Numeric vector giving either welfare or welfare share for first decile}

\item{decile2}{Numeric vector giving either welfare or welfare share for second decile}

\item{decile3}{Numeric vector giving either welfare or welfare share for third decile}

\item{decile4}{Numeric vector giving either welfare or welfare share for fourth decile}
}
\value{
Numeric vector giving the Palma ratio, the ratio of top 10\% to bottom 40\% welfare
}
\description{
Calculate Palma ratio using decile information
}
\examples{
if (suppressWarnings(require(pipr))) {
  pip <- get_stats()

  get_lh_palma_ratio(
    top10   = pip$decile10,
    decile1 = pip$decile1,
    decile2 = pip$decile2,
    decile3 = pip$decile3,
    decile4 = pip$decile4,
    ) |>
    head()

# Build the bottom 40 first

pip <- pip |>
transform(bottom40 = decile1 + decile2 + decile3 + decile4)

  get_lh_palma_ratio(
    top10    = pip$decile10,
    bottom40 = pip$bottom40
    ) |>
    head()
}
}
