% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lorenz.R
\name{get_lorenz}
\alias{get_lorenz}
\title{Lorenz curve}
\usage{
get_lorenz(
  .data,
  welfare,
  weight,
  distribution_type = c("micro", "group", "aggregate", "imputed"),
  nbins = if (length(welfare) > 1000) 100 else 20
)
}
\arguments{
\item{.data}{Household survey data frame with at least a welfare variable.}

\item{welfare}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights. Default is a vector of ones,
\code{rep(1, length(welfare))}.}

\item{distribution_type}{character: Type of distribution, either micro,
group, aggregate or imputed.}

\item{nbins}{numeric: number of points on the Lorenz curve. if \code{NULL} the
returning  Lorenz curve would be the  length of the original welfare vector
minus the number of \code{NAs} of different observations in  \code{welfare} and
\code{weight}. Default is \code{100} for \code{length(welfare) > 1000} and \code{20} otherwise.}
}
\value{
data.frame
}
\description{
Calculate points on the Lorenz curve. \strong{TO BE DOCUMENTED}.
}
\examples{
data("md_ABC_2000_income")
df <- md_ABC_2000_income
lorenz <- get_lorenz(df, welfare, weight)
str(lorenz)
}
