% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low-hanging_indicators.R
\name{get_number_poor}
\alias{get_number_poor}
\title{Calculate the number of poor using headcount and population}
\usage{
get_number_poor(headcount, pop)
}
\arguments{
\item{headcount}{A numeric vector giving the headcount ratio}

\item{pop}{A numeric vector giving population size}
}
\value{
A numeric vector giving number of poor for each observation
}
\description{
Calculate the number of poor using headcount and population
}
\examples{
if (suppressWarnings(require(pipr))) {
  pip <- get_stats()
  get_number_poor(
    pip$headcount,
    pip$pop) |>
    head()
}
}
