% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_synth_vector.R
\name{get_synth_vector}
\alias{get_synth_vector}
\title{get synthetic vector based on data level}
\usage{
get_synth_vector(dt, pop, mean, level)
}
\arguments{
\item{dt}{dataframe with group data}

\item{mean}{numeric. Named vector of welfare mean for each data level
available in data frame \code{dt}}

\item{level}{charcter: data level. itcould  be nations, urban, rural, or any
other subnational division}
}
\value{
data.frame
}
\description{
get synthetic vector based on data level
}
