% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low-hanging_indicators.R
\name{get_total_shortfall}
\alias{get_total_shortfall}
\title{Calculate total shortfall from headcount, poverty gap, and poverty line}
\usage{
get_total_shortfall(headcount, pop, povgap, povline)
}
\arguments{
\item{headcount}{Numeric vector giving headcount ratio, i.e. P0 FGT measure}

\item{pop}{Numeric vector giving population size}

\item{povgap}{Numeric vector giving poverty gap ratio, i.e. P1 FGT measure}

\item{povline}{Numeric vector giving the poverty line}
}
\value{
Numeric vector giving total shortfall in the same units as the poverty line
}
\description{
Calculate total shortfall from headcount, poverty gap, and poverty line
}
