% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weighted_median.R
\name{get_weighted_median}
\alias{get_weighted_median}
\title{Weighted median}
\usage{
get_weighted_median(x, weight)
}
\arguments{
\item{x}{Numeric vector}

\item{weight}{Numeric weight vector}
}
\value{
Weighted median of x
}
\description{
Weighted median
}
