% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grouped_data_ex1}
\alias{grouped_data_ex1}
\title{A dataset containing fake group data distribution}
\format{
A data frame with 10 rows and 2 variables:
\describe{
\item{welfare}{A vector of welfate values}
\item{weight}{A vector of population weight}
}
}
\source{
Worldbank internal.
}
\usage{
grouped_data_ex1
}
\description{
A dataset containing fake group data distribution
}
\keyword{datasets}
