% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz2.R
\name{lorenz2}
\alias{lorenz2}
\title{lorenz}
\usage{
lorenz2(welfare, weight = NULL, nbins = NULL, na.rm = FALSE)
}
\arguments{
\item{welfare}{numeric: vector of welfare measures}

\item{weight}{numeric: vector of weights}
}
\value{
data.frame
}
\description{
Given a vector of weights and welfare, this functions computes the
Lorenz curve and selveral other related indicators.
}
\examples{
data("md_ABC_2010_income")
df <- md_ABC_2010_income
lz <- lorenz2(df$welfare, df$weight)[]
}
