% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz_alt.R
\name{lorenz_alt}
\alias{lorenz_alt}
\title{lorenz_alt}
\usage{
lorenz_alt(
  welfare,
  weight,
  nbins = if (length(welfare) > 1000) 100 else 20,
  na.rm = FALSE
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights.}

\item{nbins}{numeric: number of points on the Lorenz curve.}

\item{na.rm}{logical: Set to TRUE to remove missing.}
}
\value{
data.table
}
\description{
Given a vector of weights and welfare, this functions computes the
Lorenz curve and several other related indicators.
}
\examples{
wbpip:::lorenz_alt(welfare = 1:2000, weight = rep(1, 2000))
}
\keyword{internal}
