% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_check_data.R
\name{md_check_data}
\alias{md_check_data}
\title{Check microdata to be used in PIP methods}
\usage{
md_check_data(dt, ...)
}
\arguments{
\item{dt}{Data frame.}

\item{...}{list of arguments that correspond to specific variable names.
Arguments available are in \code{getOption("wbpip.agrs_to_check")}.
For instance, welfare = "income", weight = "peso'.}
}
\value{

}
\description{
Check microdata to be used in PIP methods
}
