% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_clean_data.R
\name{md_clean_data}
\alias{md_clean_data}
\title{Clean data (micro)}
\usage{
md_clean_data(dt, welfare, weight = NULL, quiet = FALSE)
}
\arguments{
\item{dt}{data.frame: A table with survey data.}

\item{welfare}{character: Name of welfare column.}

\item{weight}{character: Name of weight column. Optional.}

\item{quiet}{logical: If TRUE output messages are suppressed.}
}
\value{
list
}
\description{
Clean microdata to be used in PIP methods.
}
\details{
\code{md_clean_data()} returns a list of elements whose main object is a
data.table with the necessary transformations to be included in PIP methods.
Data is available in element \verb{$data}. The other elements provide the number
of observations that were modified depending on test performed. The name of
elements are in the form p_s, where p (or prefix) refers to the test and s
(the suffix) refers to the name of the variable evaluated.

Prefixes are:
\itemize{
\item nna: Number of NA in variable
\item nng: Number of negative values
\item ina: Index of obs with NA in variable
\item ing: Index of obs with negative values
}
}
\keyword{internal}
