% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_quantiles.R
\name{md_compute_cumulative_share}
\alias{md_compute_cumulative_share}
\title{Welfare share by quantile in micro data}
\usage{
md_compute_cumulative_share(
  welfare = NULL,
  weight = rep(1, length = length(welfare)),
  n_quantile = 10,
  lorenz = NULL
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights. Default is a vector of ones}

\item{n_quantile}{numeric: Number of quantiles for which share of total income
is desired. Default is 10.}

\item{lorenz}{numeric: Output from \code{md_compute_lorenz}}
}
\value{
list with vector of share of welfare by quantiles
}
\description{
\code{md_welfare_share_at} returns the share of welfare held by an specified
share of the population. You can select the number of quantiles (10 be default).
This function makes use of \code{md_compute_lorenz}.
}
\examples{
md_compute_cumulative_share(welfare = md_GHI_2000_consumption$welfare,
                            weight = md_GHI_2000_consumption$weight)
}
