% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_dist_stats.R
\name{md_compute_dist_stats}
\alias{md_compute_dist_stats}
\title{Compute distributional statistics}
\usage{
md_compute_dist_stats(
  welfare,
  weight,
  mean = NULL,
  nbins = if (length(welfare) > 1000) 100 else 20,
  lorenz = NULL,
  n_quantile = 10,
  ppp_year = getOption("wbpip.available_ppp_years")
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{weight}{numeric: A vector of weights. Default is a vector of ones,
\code{rep(1, length(welfare))}.}

\item{mean}{numeric: A value with the mean. Optional.}

\item{nbins}{numeric: number of points on the Lorenz curve. if \code{NULL} the
returning  Lorenz curve would be the  length of the original welfare vector
minus the number of \code{NAs} of different observations in  \code{welfare} and
\code{weight}. Default is \code{100} for \code{length(welfare) > 1000} and \code{20} otherwise.}

\item{lorenz}{data.frame: A table with points on the Lorenz curve. Output of
\code{md_compute_lorenz()}. Optional.}

\item{n_quantile}{numeric: Number of quantiles for which share of total
income is desired. It can't be larger that the total number of percentiles
in the Lorenz curve provided by the user. Default is 10.}

\item{ppp_year}{numeric PPP year - 2017 (default) or 2011}
}
\value{
data.frame
}
\description{
Given a vector of weights and welfare, this functions computes the various
distributional statistics.
}
\examples{
wbpip:::md_compute_dist_stats(welfare = 1:2000, weight = rep(1, 2000))
}
\keyword{internal}
