% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_headcount.R
\name{md_compute_headcount}
\alias{md_compute_headcount}
\title{Compute Headcount}
\usage{
md_compute_headcount(welfare, weight, povline_lcu)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights.}

\item{povline_lcu}{numeric: Poverty line in Local Currency Unit (LCU).}
}
\value{
numeric
}
\description{
Compute headcount for microdata.
}
\examples{
wbpip:::md_compute_headcount(
  welfare = 1:2000,
  weight = rep(1, 2000),
  povline_lcu = 10
)
}
\keyword{internal}
