% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_lorenz.R
\name{md_compute_lorenz_old}
\alias{md_compute_lorenz_old}
\title{Lorenz curve}
\usage{
md_compute_lorenz_old(welfare, weight, nbins = NULL)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights.}

\item{nbins}{numeric: number of points on the Lorenz curve.}
}
\value{
data.frame
}
\description{
Compute the Lorenz curve for microdata.
}
\details{
Given a vector of weights and welfare, this functions computes the
Lorenz curve.
}
\examples{
md_compute_lorenz_old(welfare = 1:2000, weight = rep(1, 2000))
}
