% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_mld.R
\name{md_compute_mld}
\alias{md_compute_mld}
\title{Mean Log Deviation}
\usage{
md_compute_mld(welfare, weight, mean = NULL)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{weight}{numeric: A vector of weights. Default is a vector of ones,
\code{rep(1, length(welfare))}.}

\item{mean}{numeric: A value with the mean. Optional.}
}
\value{
numeric
}
\description{
Given a vector of weights and welfare, this functions computes the
Mean Log Deviation (MLD).
}
\examples{
wbpip:::md_compute_mld(welfare = 1:2000, weight = rep(1, 2000))
}
\keyword{internal}
