% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_mld_old.R
\name{md_compute_mld_old}
\alias{md_compute_mld_old}
\title{Mean Log Deviation}
\usage{
md_compute_mld_old(welfare, weight)
}
\arguments{
\item{welfare}{numeric: vector of welfare measures}

\item{weight}{numeric: vector of weights}
}
\value{
numeric
}
\description{
Given a vector of weights and welfare, this functions computes the
Mean Log Deviation (MLD).
}
\examples{
wbpip:::md_compute_mld_old(welfare = 1:2000, weight = rep(1, 2000))

}
\keyword{internal}
