% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_poverty_stats_replacement.R
\name{md_compute_pov_severity}
\alias{md_compute_pov_severity}
\title{Compute poverty severity}
\usage{
md_compute_pov_severity(
  welfare = NULL,
  weight = NULL,
  povline = NULL,
  weight_pov = NULL,
  weight_total = NULL,
  relative_distance = NULL,
  verbose = FALSE
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{weight}{numeric: A vector of population weights, optional, a vector of 1s if not specified.}

\item{povline}{numeric: Poverty line in international dollars}

\item{weight_pov}{numeric: A vector of population weights for the population
in poverty. Default is NULL, primary purpose is internal}

\item{weight_total}{numeric: sum of population weights}

\item{relative_distance}{numeric: vector of relative, standardized distances
to poverty line for the poor. Default is NULL, which means it will be computed.}

\item{verbose}{logical: display messages. Default is FALSE}
}
\value{
numeric: poverty severity
}
\description{
Compute poverty severity
}
