% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_poverty_stats.R
\name{md_compute_poverty_stats}
\alias{md_compute_poverty_stats}
\title{Compute Poverty Statistics}
\usage{
md_compute_poverty_stats(welfare, weight, povline_lcu, cons_floor = 0.5)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{povline_lcu}{numeric: Poverty line in Local Currency Unit (LCU).}

\item{cons_floor}{numeric: gives the consumption floor. Default is $0.5 per day, if NULL then no floor.}
}
\value{
list
}
\description{
Compute poverty statictics for microdata.
}
\details{
Given a vector of consumption or income values and their respective weights
\code{md_compute_poverty_stats()} computes poverty headcount, poverty gap,
poverty severity and the watts index.
}
\examples{
wbpip:::md_compute_poverty_stats(
  welfare = 1:2000,
  weight = rep(1, 2000),
  povline_lcu = 10
)
}
\keyword{internal}
