% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_prosperity_gap.R
\name{md_compute_prosperity_gap}
\alias{md_compute_prosperity_gap}
\title{Compute Prosperity Gap}
\usage{
md_compute_prosperity_gap(welfare, weight, povline_lcu, cons_floor = c(0.5))
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{cons_floor}{numeric: gives the consumption floor. Default is $0.5 per day, if NULL then no floor.}
}
\value{
numeric
}
\description{
Compute prosperity gap using microdata.
}
