% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_spl.R
\name{md_compute_spl}
\alias{md_compute_spl}
\title{Compute Societal Poverty Line (SPL)}
\usage{
md_compute_spl(
  welfare = NULL,
  weight = NULL,
  weighted_median_welfare = NULL,
  ppp_year = getOption("wbpip.available_ppp_years")
)
}
\arguments{
\item{welfare}{numeric welfare vector}

\item{weight}{numeric weight vector}

\item{weighted_median_welfare}{numeric weighted welfare median}

\item{ppp_year}{numeric PPP year - 2017 (default) or 2011}
}
\value{
numeric value giving SPL for given
}
\description{
Compute SPL based off median welfare in 2017 or 2011 PPP terms
}
