% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_poverty_stats.R,
%   R/md_compute_poverty_stats_replacement.R
\name{md_compute_watts}
\alias{md_compute_watts}
\title{Compute Watts Index on micro data}
\usage{
md_compute_watts(
  welfare = NULL,
  weight = NULL,
  povline = NULL,
  verbose = FALSE
)

md_compute_watts(
  welfare = NULL,
  weight = NULL,
  povline = NULL,
  verbose = FALSE
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{weight}{numeric: A vector of population weights, optional, a vector of 1s if not specified.}

\item{povline}{numeric: Poverty line in international dollars}

\item{verbose}{logical: display messages. Default is FALSE}
}
\value{
numeric: watts index

numeric: watts index
}
\description{
Compute Watts Index on micro data

Compute Watts Index on micro data
}
