% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microdata.R
\docType{data}
\name{md_GHI_2000_income}
\alias{md_GHI_2000_income}
\title{GHI 2000 Income (microdata)}
\format{
A data frame with 2000 rows and 6 variables:
\describe{
\item{country_code}{A fake three letter country code.}
\item{survey_year}{The year of the survey.}
\item{weight}{The weight of each indvidual's welfare value.}
\item{welfare}{The welfare of each indvidual. Can represent either income or consumption.}
\item{area}{Geographical location; urban or rural.}
\item{gender}{Gender; male or female.}
}
}
\source{
Worldbank internal.
}
\usage{
md_GHI_2000_income
}
\description{
A dataset containing synthetic microdata for 2000 individuals in the fake
country GHI for the year 2000. Income based welfare values.
}
\details{
Note that nothing in this dataset is real. It is only meant for testing and
example purposes.
}
\keyword{datasets}
