% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_quantile_functions.R
\name{md_welfare_share_at}
\alias{md_welfare_share_at}
\title{Welfare share by quantile in micro data}
\usage{
md_welfare_share_at(
  welfare = NULL,
  weight = rep(1, length = length(welfare)),
  n = 10,
  format = c("dt", "list", "atomic")
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights. Default is a vector of ones,
\code{rep(1, length(welfare))}.}
}
\value{
list with vector of share of welfare by quantiles
}
\description{
\code{md_welfare_share_at} returns the share of welfare held by an specified
share of the population. You can select the number of quantiles (10 be default).
This function makes use of \code{md_compute_lorenz}.
}
\examples{
md_welfare_share_at(welfare = md_GHI_2000_consumption$welfare,
                    weight = md_GHI_2000_consumption$weight)
}
