% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{old_gd_compute_fit_lq}
\alias{old_gd_compute_fit_lq}
\title{Computes the sum of squares of error (old version)}
\usage{
old_gd_compute_fit_lq(
  welfare,
  population,
  headcount,
  A,
  B,
  C,
  key_values = key_values
)
}
\arguments{
\item{welfare}{numeric: Welfare vector (grouped).}

\item{population}{numeric: Population vector (grouped).}

\item{headcount}{numeric: Headcount index.}

\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[3]}.}

\item{key_values}{named list: key values for lq fit, calculated using A, B, C
parameters using \link{gd_lq_key_values}}
}
\value{
list
}
\description{
Measures the fit of the model to the data.
}
\keyword{internal}
