% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{old_gd_compute_quantile_lq}
\alias{old_gd_compute_quantile_lq}
\title{Compute quantiles from Lorenz Quandratic fit (old version)}
\usage{
old_gd_compute_quantile_lq(A, B, C, n_quantile = 10, key_values)
}
\arguments{
\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[3]}.}

\item{n_quantile}{numeric: Number of quantiles to return.}

\item{key_values}{named list: key values for lq fit, calculated using A, B, C
parameters using \link{gd_lq_key_values}}
}
\value{
numeric
}
\description{
\code{old_gd_compute_quantile_lq()} computes quantiles from a Lorenz Quadratic fit.
}
\keyword{internal}
