% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_function.R
\name{old_md_compute_poverty_stats}
\alias{old_md_compute_poverty_stats}
\title{Compute Poverty Statistics}
\usage{
old_md_compute_poverty_stats(welfare, weight, povline_lcu)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{povline_lcu}{numeric: Poverty line in Local Currency Unit (LCU).}
}
\value{
list
}
\description{
Compute poverty statictics for microdata.
}
\details{
Given a vector of consumption or income values and their respective weights
\code{md_compute_poverty_stats()} computes poverty headcount, poverty gap,
poverty severity and the watts index.
}
\examples{
wbpip:::md_compute_poverty_stats(
  welfare = 1:2000,
  weight = rep(1, 2000),
  povline_lcu = 10
)
}
\keyword{internal}
