% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_quantiles.R
\name{old_md_compute_quantiles}
\alias{old_md_compute_quantiles}
\title{Compute quantiles (old version)}
\usage{
old_md_compute_quantiles(
  lwelfare,
  lweight,
  percentile,
  n_quantile = 10,
  tolerance = 0.000001
)
}
\arguments{
\item{lwelfare}{numeric: cumulative share of welfare.}

\item{lweight}{numeric: cumulative share of population.}

\item{percentile}{numeric: Monetary value each percentile.}

\item{n_quantile}{numeric: Number of quantiles for which share of total income
is desired. It can't be larger that the total number of percentiles in the
Lorenz curve provided by the user.  default is 10.}

\item{tolerance}{numeric: Tolerance parameter for \code{lorenzw >= nextQ} check.}
}
\value{
list
}
\description{
Compute quantiles for microdata.
}
\details{
Calculate share of total welfare in each quantile from Lorenz curve and
the its corresponding percentile. That is, it requires a vector with the
cumulative share of the population (\code{lweight}), a vector with cumulative
share of welfare (\code{lwelfare}), and a vector the corresponding monetary value
of each percentile (\code{percentile}).
}
\examples{
lz <- wbpip:::old_md_compute_lorenz(welfare = 1:2000, weight = rep(1, 2000))
wbpip:::old_md_compute_quantiles(
  lwelfare = lz$lorenz_welfare,
  lweight = lz$lorenz_weight,
  percentile = lz$welfare,
  n_quantile = 10
)
}
\keyword{internal}
