% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popshare_lq.R
\name{popshare_lq}
\alias{popshare_lq}
\title{Title}
\usage{
popshare_lq(A, B, C, H, mu, ct, method = 1)
}
\arguments{
\item{H}{numeric: Population share}

\item{mu}{numeric: distributional mean}

\item{ct}{list of parametric quadratic Lorenz components (from \code{get_components_lq()})}

\item{method}{numeric: estimation method. Either 1 or 2. Default 1.}
}
\value{

}
\description{
Title
}
