% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_request_year_mean.R
\name{predict_request_year_mean}
\alias{predict_request_year_mean}
\title{Predict request year mean}
\usage{
predict_request_year_mean(
  survey_year,
  survey_mean,
  proxy = list(value0, value1 = NULL, req_value)
)
}
\arguments{
\item{survey_year}{numeric: A vector with one or two survey years.}

\item{survey_mean}{numeric: A vector with one or two survey means.}

\item{proxy}{list: A list with proxy values.
\itemize{
\item \code{value0} numeric: The proxy value(s) for the first survey year.
\item \code{value1} numeric: The proxy value(s) for the second survey year.
\item \code{req_value} numeric: The proxy value for the request year.
}}
}
\value{
numeric
}
\description{
Predict a welfare mean for a request year for which survey data is not
available.
}
\details{
The survey mean(s) must be in comparable international dollars and adjusted
for differences in purchasing-power, and changes in prices and currencies.

You will also need to supply a set of \code{proxy} values to calculate the growth
rate for the distribution. Typically this is national accounts data, e.g. the
Gross Domestic Product (GDP) or Household Final Consumption Expenditure
(HFCE).

In case the survey spans two years, and you are using a decimal survey year
to take this into account, you will need to supply two proxy values for the
corresponding survey year.
}
\examples{
# Extrapolate a single survey
predict_request_year_mean(
  survey_year = 2005,
  survey_mean = 2.0,
  proxy = list(value0 = 1350, req_value = 1500))

# Interpolate two surveys (monotonic)
predict_request_year_mean(
  survey_year = c(2000, 2005),
  survey_mean = c(2.0, 3.0),
  proxy = list(value0 = 1350, value1 = 1600, req_value = 1500))

# Interpolate two surveys (non-monotonic)
predict_request_year_mean(
  survey_year = c(2000, 2005),
  survey_mean = c(2.0, 3.0),
  proxy = list(value0 = 1350, value1 = 1500, req_value = 1600))

# Extrapolate a single survey (w/ decimal year)
predict_request_year_mean(
  survey_year = 2000.3,
  survey_mean = 2.0,
  proxy = list(value0 = c(1350, 1400), req_value = 1600))

}
\references{
Prydz, E.B., D. Jolliffe, C. Lakner, D.G. Mahler, P. Sangraula. 2019.
"\href{http://documents1.worldbank.org/curated/en/664751553100573765/pdf/135460-WP-PUBLIC-Disclosed-3-21-2019.pdf}{National Accounts Data used in Global Poverty Measurement}".
Global Poverty Monitoring Technical Note 8.
World Bank, Washington, DC.
}
\seealso{
\code{\link[=deflate_welfare_mean]{deflate_welfare_mean()}} \code{\link[=fill_gaps]{fill_gaps()}}
}
