% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_gd_compute_hc_stat.R
\name{prod_gd_compute_hc_stat}
\alias{prod_gd_compute_hc_stat}
\title{Computes headcount statistic (grouped)
Version used in prod}
\usage{
prod_gd_compute_hc_stat(
  welfare,
  povline,
  population,
  requested_mean,
  popshare = NULL,
  default_ppp,
  ppp = NULL,
  p0 = 0.5
)
}
\arguments{
\item{welfare}{numeric: Cumulative proportion of welfare held by that
proportion of the population (Lorenz Curve).}

\item{povline}{numeric: Poverty line.}

\item{population}{numeric: Cumulative proportion of population.}

\item{requested_mean}{numeric: Welfare mean.}

\item{popshare}{numeric: Share of population living below the poverty line.
Optional.}

\item{default_ppp}{numeric: Default purchasing power parity.}

\item{ppp}{numeric: PPP request by user.}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}
}
\value{
list
}
\description{
Compute headcount for grouped data by selecting the best functional
fit for the Lorenz curve (either beta or quadratic).
}
\keyword{internal}
