% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_gd_compute_pip_stats.R
\name{prod_gd_compute_pip_stats}
\alias{prod_gd_compute_pip_stats}
\title{Computes poverty statistics (grouped, prod)}
\usage{
prod_gd_compute_pip_stats(
  welfare,
  povline,
  population,
  mean,
  svy_median_lcu,
  svy_median_ppp,
  popshare = NULL,
  default_ppp,
  ppp = NULL,
  p0 = 0.5
)
}
\arguments{
\item{welfare}{numeric: Cumulative proportion of welfare held by that
proportion of the population (Lorenz Curve).}

\item{povline}{numeric: Poverty line. (default 1)}

\item{population}{numeric: Cumulative proportion of population.}

\item{mean}{numeric: Welfare mean. (default 1)}

\item{popshare}{numeric: Share of population living below the poverty line.
Optional.}

\item{default_ppp}{numeric: Default purchasing power parity.}

\item{ppp}{numeric: PPP request by user.}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}
}
\value{
list
}
\description{
Compute poverty statistics for grouped data by selecting the best functional
fit for the Lorenz curve (either beta or quadratic).
}
\details{
Version used in production.
}
\keyword{internal}
