% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_gd_compute_hc_stat.R
\name{prod_gd_select_lorenz_headcount}
\alias{prod_gd_select_lorenz_headcount}
\title{Select best Lorenz fit
Version used in production}
\usage{
prod_gd_select_lorenz_headcount(lq, lb)
}
\arguments{
\item{lq}{list: Results from Lorenz Quadratic functional form. output of
\code{gd_compute_pip_stats_lq()}.}

\item{lb}{list: Results from Lorenz Beta functional form. output of
\code{gd_compute_pip_stats_lb()}.}
}
\value{
numeric
}
\description{
Select best Lorenz fit and adjust the returned statistics if needed.
}
\keyword{internal}
