% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_utils.R
\name{regres}
\alias{regres}
\title{Perform linear regression on Lorenz formatted input}
\usage{
regres(data, is_lq = TRUE)
}
\arguments{
\item{data}{list: Output of \code{create_functional_form_lq()} or
\code{create_functional_form_lb()}.}

\item{is_lq}{logical: TRUE if Lorenz Quadratic, FALSE if Beta Lorenz.}
}
\value{
list
}
\description{
\code{regres()} performs linear regression on lorenz formatted input (Beta or
Quadratic). There is no intercept in the regression. The coefficients of
regressions are estimated by ordinary least squares.
}
\examples{
# Load example data
p <- system.file("tests/testdata/gd_ex2.RDS", package = "wbpip")
df <- readRDS(p)

# Beta Lorenz
lb_data <- wbpip:::create_functional_form_lb(df$welfare, df$weight)
lb_res <- wbpip:::regres(lb_data)

# Quadratic Lorenz
lq_data <- wbpip:::create_functional_form_lq(df$welfare, df$weight)
lq_res <- wbpip:::regres(lq_data)

}
\keyword{internal}
