% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_select_lorenz.R
\name{replace_x}
\alias{replace_x}
\title{Make sure that value of x in l is valid. Otherwise use the value in o}
\usage{
replace_x(x, l, o)
}
\arguments{
\item{x}{characer: name of value in l and o}

\item{l}{list: list to use}

\item{o}{list: alternative list}
}
\value{
value
}
\description{
Make sure that value of x in l is valid. Otherwise use the value in o
}
\keyword{internal}
