% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_select_lorenz.R
\name{retrieve_distributional}
\alias{retrieve_distributional}
\title{Algorithm to retrieve correct distributional statistics}
\usage{
retrieve_distributional(lq, lb, is_valid, use_lq_for_dist)
}
\arguments{
\item{lq}{list: Results from Lorenz Quadratic functional form. output of
\code{gd_compute_pip_stats_lq()}.}

\item{lb}{list: Results from Lorenz Beta functional form. output of
\code{gd_compute_pip_stats_lb()}.}

\item{is_valid}{logical: Whether at least one of the Lorenz fit is valid}

\item{is_lq_for_dist}{logical: Whether to use LQ (TRUE) or Beta fit (FALSE)}
}
\value{
list
}
\description{
Algorithm to retrieve correct distributional statistics
}
\keyword{internal}
