% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_select_lorenz.R
\name{retrieve_poverty}
\alias{retrieve_poverty}
\title{Algorithm to retrieve correct poverty statistics}
\usage{
retrieve_poverty(lq, lb, is_normal, use_lq_for_pov)
}
\arguments{
\item{lq}{list: Results from Lorenz Quadratic functional form. output of
\code{gd_compute_pip_stats_lq()}.}

\item{lb}{list: Results from Lorenz Beta functional form. output of
\code{gd_compute_pip_stats_lb()}.}

\item{is_normal}{logical: Whether at least one of the Lorenz fit is normal}

\item{is_lq_for_pov}{logical: Whether to use LQ (TRUE) or Beta fit (FALSE)}
}
\value{
list
}
\description{
Algorithm to retrieve correct poverty statistics
}
\keyword{internal}
