% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_utils.R
\name{return_format}
\alias{return_format}
\title{Return data according to format}
\usage{
return_format(ld, var, complete = FALSE, format = c("dt", "list", "atomic"))
}
\arguments{
\item{ld}{list of data}

\item{var}{character: name of variable to be returned.}

\item{format}{character: either "dt" for data.table, "list" or "atomic" for a
single numeric vector, whose names are corresponding selected Lorenz for
each value.  Default is "dt"}
}
\value{
data.table, list, or atomic vector
}
\description{
Return data according to format
}
\keyword{internal}
