% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{rtNewt}
\alias{rtNewt}
\title{rtNewt}
\usage{
rtNewt(mean, povline, A, B, C)
}
\arguments{
\item{mean}{numeric: Welfare measure mean (income of consumption)}

\item{povline}{numeric: Poverty line}

\item{A}{numeric vector: lorenz curve coefficient}

\item{B}{numeric vector: lorenz curve coefficient}

\item{C}{numeric vector: lorenz curve coefficient}
}
\value{
numeric
}
\description{
rtNewt
}
