% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{rtSafe}
\alias{rtSafe}
\title{rtSafe}
\usage{
rtSafe(x1, x2, xacc, mean, povline, A, B, C)
}
\arguments{
\item{x1}{numeric}

\item{x2}{numeric}

\item{xacc}{numeric}

\item{mean}{numeric: Welfare measure mean (income of consumption)}

\item{povline}{numeric: Poverty line}

\item{A}{numeric vector: lorenz curve coefficient}

\item{B}{numeric vector: lorenz curve coefficient}

\item{C}{numeric vector: lorenz curve coefficient}
}
\value{
numeric
}
\description{
rtSafe
}
\examples{
rtSafe(x1, x2, xacc, povline, mean, A, B, C)

}
