% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_compute_dist_stats.R
\name{st_create_synth_vector}
\alias{st_create_synth_vector}
\title{Create synthetic vector based on Lorenz  parameters of group data}
\usage{
st_create_synth_vector(welfare, population, mean, pop, p0 = 0.5)
}
\arguments{
\item{welfare}{numeric: Cumulative proportion of welfare held by that
proportion of the population (Lorenz Curve).}

\item{population}{numeric: Cumulative proportion of population.}

\item{mean}{numeric: Welfare mean.}

\item{pop}{numeric: a sinble number with the population at data_level}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}
}
\value{
list
}
\description{
Compute distributional statistics for grouped data by selecting the best
functional fit for the Lorenz curve (either beta or quadratic).
}
\keyword{internal}
