% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_clean_data.R
\name{standardize_type2}
\alias{standardize_type2}
\title{standardize_type2
Standardize grouped data of type 2 distribution}
\usage{
standardize_type2(population, welfare)
}
\arguments{
\item{population}{numeric: population, proportion of population}

\item{welfare}{numeric: welfare, proportion of income / consumption}
}
\value{
list
}
\description{
standardize_type2
Standardize grouped data of type 2 distribution
}
