% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_clean_data.R
\name{standardize_type5}
\alias{standardize_type5}
\title{Standardize type 5}
\usage{
standardize_type5(welfare, population)
}
\arguments{
\item{population}{}
}
\value{

}
\description{
Standardize type 5
}
