% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_clean_data.R
\name{standardize_type5_old}
\alias{standardize_type5_old}
\title{standardize_type5_old
Standardize grouped data of type 5 distribution}
\usage{
standardize_type5_old(
  population,
  welfare,
  min_welfare_default,
  max_welfare_default
)
}
\arguments{
\item{population}{numeric: population, in percentage of the population in a given interval of incomes.}

\item{welfare}{numeric: welfare, the mean income / consumption of that interval}

\item{min_welfare_default}{numeric:}

\item{max_welfare_default}{numeric:}
}
\value{
list
}
\description{
standardize_type5_old
Standardize grouped data of type 5 distribution
}
