% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_select_lorenz.R
\name{use_lq_for_distributional}
\alias{use_lq_for_distributional}
\title{Algorithm to decide which Lorenz fit to use for distributional statistics}
\usage{
use_lq_for_distributional(lq, lb)
}
\arguments{
\item{lq}{list: Results from Lorenz Quadratic functional form. output of
\code{gd_compute_pip_stats_lq()}.}

\item{lb}{list: Results from Lorenz Beta functional form. output of
\code{gd_compute_pip_stats_lb()}.}
}
\value{
logical:
returns TRUE for Lorenz Quadratic
returns FALSE for Lorenz Beta
}
\description{
Algorithm to decide which Lorenz fit to use for distributional statistics
}
\keyword{internal}
