% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{value_at_lb}
\alias{value_at_lb}
\title{Solves for beta Lorenz curves}
\usage{
value_at_lb(x, A, B, C)
}
\arguments{
\item{x}{numeric: Point on curve.}

\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[3]}.}
}
\value{
numeric
}
\description{
\code{value_at_lb()}solves for beta Lorenz curves.
}
\keyword{internal}
