% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_dist_stats.R
\name{check_curve_validity_dist_lb}
\alias{check_curve_validity_dist_lb}
\title{check_curve_validity_dist_lb
Check validity of Lorenz beta fit for distributional statistics}
\usage{
check_curve_validity_dist_lb(A, B, C)
}
\arguments{
\item{A}{numeric: First regression coefficient.}

\item{B}{numeric: Second regression coefficient.}

\item{C}{numeric: Third regression coefficient.}
}
\value{
list
}
\description{
check_curve_validity_dist_lb
Check validity of Lorenz beta fit for distributional statistics
}
\keyword{internal}
