% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{check_curve_validity_lb}
\alias{check_curve_validity_lb}
\title{Check validity of lorenz beta fit}
\usage{
check_curve_validity_lb(A, B, C)
}
\arguments{
\item{A}{numeric: First regression coefficient}

\item{B}{numeric: Second regression coefficient}

\item{C}{numeric: Third regression coefficient}
}
\value{
list
}
\description{
\code{check_curve_validity_lb()} checks the validity of the lorenz beta fit
}
\seealso{
\href{https://econpapers.repec.org/article/ecmemetrp/v_3a48_3ay_3a1980_3ai_3a2_3ap_3a437-46.htm}{Original Beta Lorenz curve paper}

\href{https://www.ifpri.org/cdmref/p15738coll2/id/125673}{
Computational Tools For Poverty Measurement And Analysis}
}
