% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R, R/gd_estimate_lq.R
\name{check_curve_validity_lq}
\alias{check_curve_validity_lq}
\title{Check validity of Lorenz Quadratic fit}
\usage{
check_curve_validity_lq(A, B, C, e, m, n, r)

check_curve_validity_lq(A, B, C, e, m, n, r)
}
\arguments{
\item{A}{numeric: First regression coefficient}

\item{B}{numeric: Second regression coefficient}

\item{C}{numeric: Third regression coefficient}

\item{e}{numeric: e = -(A + B + C + 1): condition for the curve to go through
(1, 1)}

\item{m}{numeric: m = (B^2) - (4 * A). m < 0: condition for the curve to be
an ellipse (m is called alpha in paper)}

\item{n}{numeric: n = (2 * B * e) - (4 * C). n is called Beta in paper}

\item{r}{r = (n^2) - (4 * m * e^2). r is called K in paper}
}
\value{
list

list
}
\description{
\code{check_curve_validity_lq()} checks the validity of the Lorenz Quadratic fit

\code{check_curve_validity_lq()} checks the validity of the Lorenz Quadratic fit
}
\references{
Datt, G. 1998. "\href{https://www.ifpri.org/cdmref/p15738coll2/id/125673}{Computational Tools For Poverty Measurement And Analysis}". FCND
Discussion Paper 50. World Bank, Washington, DC.

Krause, M. 2013. "\href{https://doi.org/10.1016/j.jeconom.2013.01.001}{Corrigendum to Elliptical Lorenz curves}". \emph{Journal of
Econometrics 174} (1): 44.

Villasenor, J., B. C. Arnold. 1989. "\href{https://EconPapers.repec.org/RePEc:eee:econom:v:40:y:1989:i:2:p:327-338}{Elliptical Lorenz curves}".
\emph{Journal of Econometrics 40} (2): 327-338.
}
\seealso{
\href{https://EconPapers.repec.org/RePEc:eee:econom:v:40:y:1989:i:2:p:327-338}{Original quadratic Lorenz curve paper}

\href{https://www.sciencedirect.com/science/article/abs/pii/S0304407613000158?via\%3Dihub}{Corrigendum to Elliptical Lorenz Curves}

\href{https://www.ifpri.org/cdmref/p15738coll2/id/125673}{
Computational Tools For Poverty Measurement And Analysis}
}
\keyword{internal}
