% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_NA_Inf_values}
\alias{check_NA_Inf_values}
\title{Check for NA values in vector}
\usage{
check_NA_Inf_values(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
A message if any values are NA or Inf
}
\description{
It is used for now on Lorenz curves equation
}
\keyword{internal}
