% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pip_stats.R
\name{compute_pip_stats}
\alias{compute_pip_stats}
\title{Compute default PIP statistics (microdata)}
\usage{
compute_pip_stats(
  welfare,
  povline,
  population = NULL,
  requested_mean = NULL,
  popshare = NULL,
  default_ppp = 1,
  ppp = NULL,
  p0 = 0.5,
  cons_floor = c(0.5),
  distribution_type = c("micro", "group", "aggregate", "imputed"),
  ppp_year = getOption("wbpip.available_ppp_years")
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{povline}{numeric: Poverty line in international dollars}

\item{population}{numeric: A vector of population weights, optional, a vector
of 1s if not specified.}

\item{requested_mean}{numeric: Welfare mean in international dollars}

\item{popshare}{numeric: Share of population for which the corresponding
quantile is desired. Default .5 (i.e., weighted median)}

\item{default_ppp}{numeric: Default purchasing power parity}

\item{ppp}{numeric: PPP requested by user}

\item{p0}{numeric: TO be documented}

\item{cons_floor}{numeric: gives the consumption floor. Default is $0.5 per day, if NULL then no floor.}

\item{distribution_type}{character: Type of distribution, either micro,
group, aggregate or imputed.}

\item{ppp_year}{numeric: PPP year - 2017 (default) or 2011}
}
\value{
list
}
\description{
Compute poverty and distributional statistics for microdata.
}
