% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_spl.R
\name{compute_spl}
\alias{compute_spl}
\title{Compute SPL}
\usage{
compute_spl(weighted_median_ppp, ppp_year)
}
\arguments{
\item{weighted_median_ppp}{numeric: Median. Weighted median in PPP terms.}

\item{ppp_year}{numeric: PPP year.}
}
\value{
numeric vector with spl values
}
\description{
Compute the societal poverty line based on the median welfare in PPP-adjusted
dollars.
}
\references{
D. Jolliffe, E. B. Prydz. 2017. "\href{https://documents1.worldbank.org/curated/en/133671495562984832/pdf/Societal-poverty-a-relative-and-relevant-measure.pdf}{Societal Poverty: A Relative and Relevant Measure}".
Policy Research Working Paper 8073. World Bank, Washington, DC.
}
\keyword{internal}
