% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_spl.R
\name{compute_spl}
\alias{compute_spl}
\title{Compute societal poverty lines}
\usage{
compute_spl(x, ...)
}
\arguments{
\item{x}{data frame with either micro data, group data, or imputed data}

\item{welfare}{welfare vector}

\item{weight}{population weights vector}
}
\value{
scalar with poverty line
}
\description{
Compute societal poverty lines
}
