% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_spl.R
\name{compute_spl}
\alias{compute_spl}
\title{Compute societal poverty lines}
\usage{
compute_spl(dt, ...)
}
\arguments{
\item{dt}{data frame with either micro data, group data, or imputed data}

\item{welfare}{character: name of variable with welfare vector}

\item{weight}{character: name of variable with population (weight) vector}
}
\value{
scalar with poverty line
}
\description{
Compute societal poverty lines
}
