% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{create_functional_form_lb}
\alias{create_functional_form_lb}
\title{Prepares data for lorenz beta regression}
\usage{
create_functional_form_lb(welfare, population)
}
\arguments{
\item{welfare}{numeric: Welfare vector from empirical Lorenz curve}

\item{population}{numeric: Population vector from empirical Lorenz curve}
}
\value{
data.frame
}
\description{
Prepares data for regression. The last observation of (p,l), which by construction has the value
(1, 1), is excluded since the functional form for the Lorenz curve already
forces it to pass through the point (1, 1).
}
\seealso{
\href{https://econpapers.repec.org/article/ecmemetrp/v_3a48_3ay_3a1980_3ai_3a2_3ap_3a437-46.htm}{Original Beta Lorenz curve paper}
}
