% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{create_functional_form_lq}
\alias{create_functional_form_lq}
\title{Prepares data for Lorenz Quadratic regression}
\usage{
create_functional_form_lq(welfare, population)
}
\arguments{
\item{welfare}{numeric: Welfare vector from empirical Lorenz curve}

\item{population}{numeric: Population vector from empirical Lorenz curve}
}
\value{
data.frame
}
\description{
Prepares data for regression of L(1-L) on (P^2-L), L(P-1) and
(P-L). The last observation of (p,l), which by construction has the value
(1, 1), is excluded since the functional form for the Lorenz curve already
forces it to pass through the point (1, 1). Equation 15 in Lorenz Quadratic
original paper.
Then it performs linear regression on lorenz formatted input (Beta or
Quadratic). There is no intercept in the regression. The coefficients of
regressions are estimated by ordinary least squares.
}
\seealso{
\href{https://EconPapers.repec.org/RePEc:eee:econom:v:40:y:1989:i:2:p:327-338}{Original quadratic Lorenz curve paper}

\href{https://www.sciencedirect.com/science/article/abs/pii/S0304407613000158?via\%3Dihub}{Corrigendum to Elliptical Lorenz Curves}
}
