% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{derive_lb}
\alias{derive_lb}
\title{Returns the first derivative of the beta Lorenz- Vectorized}
\usage{
derive_lb(x, A, B, C)
}
\arguments{
\item{x}{numeric: Point on curve. Allow for vectors.}

\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[3]}.}
}
\value{
numeric
}
\description{
\code{derive_lb()} returns the first derivative of a beta Lorenz curve.
}
\keyword{internal}
