% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{derive_lb}
\alias{derive_lb}
\title{Returns the first derivative of the beta Lorenz}
\usage{
derive_lb(x, A, B, C)
}
\arguments{
\item{x}{numeric: point on curve}

\item{A}{numeric vector: lorenz curve coefficient}

\item{B}{numeric vector: lorenz curve coefficient}

\item{C}{numeric vector: lorenz curve coefficient}
}
\value{
numeric
}
\description{
\code{derive_lb()} returns the first derivative of the beta Lorenz curves
}
