% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{derive_lq}
\alias{derive_lq}
\title{Returns the first derivative of the Lorenz quadratic}
\usage{
derive_lq(x, ct, method = 1)
}
\arguments{
\item{x}{numeric: point on curve (i.e., share of population)}

\item{ct}{list: components from \code{get_components_lq}}

\item{method}{numeric: Method 1 is the share of mean that corresponds to the
threshold that produces the defiles share of population}
}
\value{
numeric
}
\description{
\code{derive_lq()} returns the first derivative of the quadratic Lorenz curves
with c = 1. General quadratic form: ax^2 + bxy + cy^2 + dx + ey + f = 0
}
