% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{derive_lq}
\alias{derive_lq}
\title{Returns the first derivative of the quadratic Lorenz}
\usage{
derive_lq(x, A, B, C)
}
\arguments{
\item{x}{numeric: point on curve}

\item{A}{numeric vector: lorenz curve coefficient}

\item{B}{numeric vector: lorenz curve coefficient}

\item{C}{numeric vector: lorenz curve coefficient}
}
\value{
numeric
}
\description{
\code{derive_lq()} returns the first derivative of the quadratic Lorenz curves
with c = 1. General quadratic form: ax^2 + bxy + cy^2 + dx + ey + f = 0
}
