% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_gaps.R
\name{fill_gaps}
\alias{fill_gaps}
\title{Fill gaps}
\usage{
fill_gaps(
  request_year,
  data = list(df0, df1 = NULL),
  predicted_request_mean,
  survey_year,
  data_type = c("microdata", "groupdata"),
  poverty_line = 1.9
)
}
\arguments{
\item{request_year}{integer: A value with the request year.}

\item{data}{list: A list with one or two data frames containing survey data.
See details.}

\item{predicted_request_mean}{numeric: A vector with one or two survey means.
See details.}

\item{survey_year}{numeric: A vector with one or two survey years.}

\item{data_type}{character: Type of data, either micro or grouped.}

\item{poverty_line}{numeric: Daily poverty line in international dollars.}
}
\description{
Calculate poverty statistics for a request year for which survey data is not
available.
}
\details{
The predicted request year mean(s) must be in comparable international
dollars and adjusted for differences in purchasing-power, and changes in
prices and currencies.

The survey data must contain a column named \emph{welfare} and optionally a column
named \emph{weight} if welfare values are to be weighted.
}
\examples{
# Load example data
data('md_ABC_2000_income')
data('md_ABC_2010_income')
md_ABC_2010_income <-
   md_clean_data(md_ABC_2010_income,
     welfare = 'welfare',
     weight = 'weight')$data

# Extrapolation
res <- fill_gaps(
  request_year = 2005,
  survey_year = 2000,
  data = list(df0 = md_ABC_2000_income),
  predicted_request_mean = 13,
  data_type = 'microdata',
  poverty_line = 1.9)

# Interpolatation (monotonic)
res <- fill_gaps(
  request_year = 2005,
  survey_year = c(2000, 2010),
  data = list(df0 = md_ABC_2000_income, df1 = md_ABC_2010_income),
  predicted_request_mean = c(13, 13),
  data_type = 'microdata',
  poverty_line = 1.9)

# Interpolatation (non-monotonic)
res <- fill_gaps(
  request_year = 2005,
  survey_year = c(2000, 2010),
  data = list(df0 = md_ABC_2000_income, df1 = md_ABC_2010_income),
  predicted_request_mean = c(14, 17),
  data_type = 'microdata',
  poverty_line = 1.9)

}
\seealso{
\code{\link[=deflate_welfare_mean]{deflate_welfare_mean()}} \code{\link[=predict_request_year_mean]{predict_request_year_mean()}}
}
