% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_dist_stats.R
\name{gd_compute_dist_fit_lq}
\alias{gd_compute_dist_fit_lq}
\title{Computes the sum of squares of error}
\usage{
gd_compute_dist_fit_lq(welfare, population, A, B, C)
}
\arguments{
\item{welfare}{numeric: Welfare vector (grouped).}

\item{population}{numeric: Population vector (grouped).}

\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[3]}.}
}
\value{
list
}
\description{
Measures the fit of the model to the data for distributional statistics.
}
\keyword{internal}
