% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_dist_stats_lq}
\alias{gd_compute_dist_stats_lq}
\title{Computes distributional stats from Lorenz Quadratic fit}
\usage{
gd_compute_dist_stats_lq(mean, p0, ct)
}
\arguments{
\item{mean}{numeric: welfare mean}

\item{p0}{numeric: To document}

\item{ct}{list: components from \code{get_components_lq}.}
}
\value{
list
}
\description{
Computes distributional stats from Lorenz Quadratic fit
}
