% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R, R/gd_estimate_lq.R
\name{gd_compute_fit_lq}
\alias{gd_compute_fit_lq}
\title{Computes the sum of squares of error}
\usage{
gd_compute_fit_lq(welfare, population, headcount, A, B, C)

gd_compute_fit_lq(welfare, population, headcount, A, B, C)
}
\arguments{
\item{welfare}{numeric: Welfare vector (grouped)}

\item{population}{numeric: Population vector (grouped)}

\item{headcount}{numeric: headcount index}

\item{A}{numeric vector: Lorenz curve coefficient}

\item{B}{numeric vector: Lorenz curve coefficient}

\item{C}{numeric vector: Lorenz curve coefficient}
}
\value{
list

list
}
\description{
Measures the fit of the model to the data.
}
\keyword{internal}
