% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_gini_lq}
\alias{gd_compute_gini_lq}
\title{Compute gini index from Lorenz Quadratic fit}
\usage{
gd_compute_gini_lq(A, B, C)
}
\arguments{
\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[3]}.}
}
\value{
numeric
}
\description{
\code{gd_compute_gini_lq()} computes the gini index from a Lorenz Quadratic fit
}
\references{
Datt, G. 1998. "\href{https://www.ifpri.org/cdmref/p15738coll2/id/125673}{Computational Tools For Poverty Measurement And Analysis}". FCND
Discussion Paper 50. World Bank, Washington, DC.
}
\keyword{internal}
