% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_gini_lq}
\alias{gd_compute_gini_lq}
\title{Compute gini index from Lorenz Quadratic fit}
\usage{
gd_compute_gini_lq(ct)
}
\arguments{
\item{ct}{list: components from \code{get_components_lq}.}
}
\value{
numeric
}
\description{
\code{gd_compute_gini_lq()} computes the gini index from a Lorenz Quadratic fit
}
\seealso{
\href{https://www.ifpri.org/cdmref/p15738coll2/id/125673}{
Computational Tools For Poverty Measurement And Analysis}
}
