% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_gini_lq}
\alias{gd_compute_gini_lq}
\title{Compute gini index from Lorenz Quadratic fit}
\usage{
gd_compute_gini_lq(A, B, C, e, m, n, r)
}
\arguments{
\item{A}{numeric: First regression coefficient}

\item{B}{numeric: Second regression coefficient}

\item{C}{numeric: Third regression coefficient}

\item{e}{numeric: e = -(A + B + C + 1): condition for the curve to go through
(1, 1)}

\item{m}{numeric: m = (B^2) - (4 * A). m < 0: condition for the curve to be
an ellipse (m is called alpha in paper)}

\item{n}{numeric: n = (2 * B * e) - (4 * C). n is called Beta in paper}

\item{r}{numeric: r = (n^2) - (4 * m * e^2). r is called K in paper}
}
\value{
numeric
}
\description{
\code{gd_compute_gini_lq()} computes the gini index from a Lorenz Quadratic fit
}
\seealso{
\href{https://www.ifpri.org/cdmref/p15738coll2/id/125673}{
Computational Tools For Poverty Measurement And Analysis}
}
