% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R,
%   R/gd_compute_pip_stats_lb2.R
\name{gd_compute_headcount_lb}
\alias{gd_compute_headcount_lb}
\title{Compute the headcount statistic from Lorenz Beta fit}
\usage{
gd_compute_headcount_lb(mean, povline, A, B, C)

gd_compute_headcount_lb(mean, povline, A, B, C)
}
\arguments{
\item{mean}{numeric: Welfare measure mean (income of consumption).}

\item{povline}{numeric: Poverty line.}

\item{A}{numeric: Lorenz curve coefficient.}

\item{B}{numeric: Lorenz curve coefficient.}

\item{C}{numeric: Lorenz curve coefficient.}
}
\value{
numeric

numeric
}
\description{
Compute the headcount statistic from Lorenz Beta fit

Compute the headcount statistic from Lorenz Beta fit
}
\keyword{internal}
