% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_headcount_lq}
\alias{gd_compute_headcount_lq}
\title{Compute poverty for Quadratic Lorenz}
\usage{
gd_compute_headcount_lq(mean, povline, B, m, n, r)
}
\arguments{
\item{mean}{numeric: Welfare mean.}

\item{povline}{numeric: Poverty line.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{m}{numeric: m = (B^2) - (4 * A). m < 0: condition for the curve to be
an ellipse (m is called alpha in paper).}

\item{n}{numeric: n = (2 * B * e) - (4 * C). n is called Beta in paper}

\item{r}{r = (n^2) - (4 * m * e^2). r is called K in paper.}
}
\value{
poverty headcount
}
\description{
Compute poverty for Quadratic Lorenz
}
\keyword{internal}
