% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_headcount_lq}
\alias{gd_compute_headcount_lq}
\title{Computes Poverty headcount for group data (Lorenz Quadratic)}
\usage{
gd_compute_headcount_lq(mean, povline, A, B, C)
}
\arguments{
\item{mean}{numeric: Welfare mean.}

\item{povline}{numeric: Poverty line.}

\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[3]}.}
}
\value{
numeric vector length 1
}
\description{
Computes Poverty headcount for group data (Lorenz Quadratic)
}
