% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_headcount_lq}
\alias{gd_compute_headcount_lq}
\title{Compute poverty for Quadratic Lorenz}
\usage{
gd_compute_headcount_lq(mean, povline, B, key_values)
}
\arguments{
\item{mean}{numeric: Welfare mean.}

\item{povline}{numeric: Poverty line.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{key_values}{named list: key values for lq fit, calculated using A, B, C
parameters using \link{gd_lq_key_values}}
}
\value{
poverty headcount
}
\description{
Compute poverty for Quadratic Lorenz
}
