% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_mld_lq}
\alias{gd_compute_mld_lq}
\title{Computes MLD from Lorenz Quadratic fit}
\usage{
gd_compute_mld_lq(dd, A, B, C)
}
\arguments{
\item{dd}{numeric: \strong{TO BE DOCUMENTED}.}

\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[3]}.}
}
\value{
numeric
}
\description{
\code{gd_compute_mld_lq()} computes the Mean Log deviation (MLD) from a Lorenz
Quadratic fit
}
