% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R, R/gd_estimate_lq.R
\name{gd_compute_mld_lq}
\alias{gd_compute_mld_lq}
\title{Computes MLD from Lorenz Quadratic fit}
\usage{
gd_compute_mld_lq(dd, A, B, C)

gd_compute_mld_lq(dd, A, B, C)
}
\arguments{
\item{dd}{numeric}

\item{A}{numeric vector: lorenz curve coefficient}

\item{B}{numeric vector: lorenz curve coefficient}

\item{C}{numeric vector: lorenz curve coefficient}
}
\value{
numeric

numeric
}
\description{
\code{gd_compute_mld_lq()} computes the Mean Log deviation (MLD) from a Lorenz
Quadratic fit

\code{gd_compute_mld_lq()} computes the Mean Log deviation (MLD) from a Lorenz
Quadratic fit
}
\keyword{internal}
