% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_mld_lq}
\alias{gd_compute_mld_lq}
\title{Computes MLD from Lorenz Quadratic fit}
\usage{
gd_compute_mld_lq(ct)
}
\arguments{
\item{ct}{list}
}
\value{
numeric
}
\description{
\code{gd_compute_mld_lq()} computes the Mean Log deviation (MLD) from a Lorenz
Quadratic fit
}
