% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats.R
\name{gd_compute_pip_stats}
\alias{gd_compute_pip_stats}
\title{Computes poverty and inequality statistics (grouped)}
\usage{
gd_compute_pip_stats(
  welfare,
  povline,
  population,
  requested_mean,
  popshare = NULL,
  default_ppp = 1,
  ppp = NULL,
  p0 = 0.5,
  ppp_year = getOption("wbpip.available_ppp_years")
)
}
\arguments{
\item{welfare}{numeric: Cumulative proportion of welfare held by that
proportion of the population (Lorenz Curve).}

\item{povline}{numeric: Poverty line.}

\item{population}{numeric: Cumulative proportion of population.}

\item{requested_mean}{numeric: Welfare mean.}

\item{popshare}{numeric: Share of population living below the poverty line.
Optional.}

\item{default_ppp}{numeric: Default purchasing power parity.}

\item{ppp}{numeric: PPP request by user.}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}

\item{ppp_year}{numeric PPP year - 2017 (default) or 2011}
}
\value{
list
}
\description{
Compute poverty statistics for grouped data by selecting the best functional
fit for the Lorenz curve (either beta or quadratic).
}
\examples{
# Compute PIP stats
res <- wbpip:::gd_compute_pip_stats(
         grouped_data_ex2$welfare,
         grouped_data_ex2$weight,
         requested_mean = 2.911786,
         povline = 1.9,
         default_ppp = 1)

}
\keyword{internal}
