% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_estimate_lq.R
\name{gd_compute_pip_stats_lq}
\alias{gd_compute_pip_stats_lq}
\title{Computes poverty statistics from grouped data}
\usage{
gd_compute_pip_stats_lq(
  grouped_data,
  mean,
  povline,
  ppp = NULL,
  default_ppp,
  popshare = NULL,
  isLQ = TRUE
)
}
\arguments{
\item{grouped_data}{dataframe: grouped data}

\item{mean}{numeric: Welfare mean}

\item{povline}{numeric: Poverty line}

\item{ppp}{numeric: PPP request by user}

\item{default_ppp}{numeric: Default purchasing power parity}

\item{popshare}{numeric: Share of population living below the poverty line.
Optional}

\item{isLQ}{boolean: indicates whether the estimation of lorenz quadratic}
}
\value{
list
}
\description{
Computes poverty statistics from grouped data
}
\examples{
gd_estimate_lq(grouped_data, , povline, default_ppp, ppp, popshare, p, l, isLQ)
}
