% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_pip_stats_lq}
\alias{gd_compute_pip_stats_lq}
\title{Computes poverty statistics from grouped data}
\usage{
gd_compute_pip_stats_lq(
  population,
  welfare,
  mean,
  default_ppp,
  type = 1,
  povline = NULL,
  ppp = NULL,
  popshare = NULL,
  is_lq = TRUE,
  p0 = 0.5
)
}
\arguments{
\item{population}{numeric: population vector whose form depends on on \code{type}.}

\item{welfare}{numeric: welfare vector whose form depends on on \code{type}.}

\item{mean}{numeric: Welfare mean if \code{type = 1} or \code{type = 2}}

\item{default_ppp}{numeric: Default purchasing power parity}

\item{type}{numeric: Type of data. if \code{type = 1}, \code{population} must be the
cumulative proportion of population and \code{welfare} must be the cumulative
proportion of income held by that proportion of the population (Lorenz Curve).
if \code{type = 2}, \code{population} must be the proportion of population and
\code{welfare} must be the proportion of income.
If \code{type = 5}, then \code{population} must be the Percentage
of the population in a given interval of incomes, whereas \code{welfare} must be
the mean income of that interval. Default is 1.}

\item{povline}{numeric: Poverty line}

\item{ppp}{numeric: PPP request by user}

\item{popshare}{numeric: Share of population living below the poverty line.
Optional}

\item{is_lq}{boolean: indicates whether the estimation of lorenz quadratic}
}
\value{
list
}
\description{
Computes poverty statistics from grouped data
}
\examples{
# gd_estimate_lq(.data, , povline, default_ppp, ppp, popshare, p, l, is_lq)
}
