% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_pip_stats_lq}
\alias{gd_compute_pip_stats_lq}
\title{Computes poverty statistics from grouped data}
\usage{
gd_compute_pip_stats_lq(
  .data,
  mean,
  povline,
  ppp = NULL,
  default_ppp,
  popshare = NULL
)
}
\arguments{
\item{.data}{dataframe: grouped data}

\item{mean}{numeric: Welfare mean}

\item{povline}{numeric: Poverty line}

\item{ppp}{numeric: PPP request by user}

\item{default_ppp}{numeric: Default purchasing power parity}

\item{popshare}{numeric: Share of population living below the poverty line.
Optional}
}
\value{
list
}
\description{
Computes poverty statistics from grouped data
}
