% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_pip_stats_lq}
\alias{gd_compute_pip_stats_lq}
\title{Computes poverty statistics from grouped data}
\usage{
gd_compute_pip_stats_lq(
  population,
  welfare,
  mean,
  povline = NULL,
  popshare = NULL,
  default_ppp = NULL,
  ppp = NULL,
  is_lq = TRUE,
  p0 = 0.5
)
}
\arguments{
\item{population}{numeric: cumulative proportion of population.}

\item{welfare}{numeric: cumulative proportion of income held by that
proportion of the population (Lorenz Curve).}

\item{mean}{numeric: Welfare mean}

\item{povline}{numeric: Poverty line}

\item{popshare}{numeric: Share of population living below the poverty line.
Optional}

\item{default_ppp}{numeric: Default purchasing power parity}

\item{ppp}{numeric: PPP request by user}

\item{is_lq}{boolean: indicates whether the estimation of lorenz quadratic}
}
\value{
list
}
\description{
Computes poverty statistics from grouped data
}
\examples{
L <- c(0.00208, 0.01013, 0.03122, 0.07083, 0.12808, 0.23498, 0.34887, 0.51994, 0.6427, 0.79201, 0.86966, 0.91277, 1)
P <- c(0.0092, 0.0339, 0.085, 0.164, 0.2609, 0.4133, 0.5497, 0.7196, 0.8196, 0.9174, 0.957, 0.9751, 1)
mu  <- 109.9 # mean
z   <- 89    # poverty line
gd_compute_pip_stats_lq(P, L, mu, z)

res <- gd_compute_pip_stats_lq(P, L, mu, z)
res$headcount
res2 <- gd_compute_pip_stats_lq(P, L, mu, popshare = res$headcount)
res2$povline
}
