% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{gd_compute_polarization_lb}
\alias{gd_compute_polarization_lb}
\title{Computes polarization index from parametric Lorenz fit}
\usage{
gd_compute_polarization_lb(mean, p0, dcm, A, B, C)
}
\arguments{
\item{mean}{numeric: Welfare mean.}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}}

\item{dcm}{numeric: \strong{TO BE DOCUMENTED}}

\item{A}{numeric: Lorenz curve coefficient.}

\item{B}{numeric: Lorenz curve coefficient.}

\item{C}{numeric: Lorenz curve coefficient.}
}
\value{
numeric
}
\description{
Used for grouped data computations,
}
\keyword{internal}
