% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_polarization_lq}
\alias{gd_compute_polarization_lq}
\title{Computes polarization index from parametric Lorenz fit}
\usage{
gd_compute_polarization_lq(mean, p0, dcm, A, B, C, key_values)
}
\arguments{
\item{mean}{numeric: Welfare mean.}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}

\item{dcm}{numeric: Distribution corrected mean}

\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[3]}.}

\item{key_values}{named list: key values for lq fit, calculated using A, B, C
parameters using \link{gd_lq_key_values}}
}
\value{
numeric
}
\description{
Used for grouped data computations
}
