% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_polarization_lq}
\alias{gd_compute_polarization_lq}
\title{Computes polarization index from Quadratic Lorenz fit}
\usage{
gd_compute_polarization_lq(mean, p0, dcm, ct)
}
\arguments{
\item{mean}{numeric: Welfare mean}

\item{p0}{numeric: To document}

\item{dcm}{numeric: To document}

\item{ct}{list: components from \code{get_components_lq}.}
}
\value{
numeric
}
\description{
Computes polarization index from Quadratic Lorenz fit
}
