% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{gd_compute_pov_gap_lb}
\alias{gd_compute_pov_gap_lb}
\title{Compute poverty gap for Lorenz Beta fit}
\usage{
gd_compute_pov_gap_lb(u, headcount, A, B, C)
}
\arguments{
\item{u}{numeric: Normalized mean.}

\item{headcount}{numeric: Headcount.}

\item{A}{numeric: First regression parameter.}

\item{B}{numeric: Second regression parameter.}

\item{C}{numeric: Third regression parameter.}
}
\value{
numeric
}
\description{
Compute poverty gap for Lorenz Beta fit
}
\keyword{internal}
