% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{gd_compute_pov_severity_lb}
\alias{gd_compute_pov_severity_lb}
\title{Compute poverty severity for Lorenz Beta fit}
\usage{
gd_compute_pov_severity_lb(u, headcount, pov_gap, A, B, C)
}
\arguments{
\item{u}{numeric: Mean? \strong{TO BE DOCUMENTED}.}

\item{headcount}{numeric: Headcount.}

\item{pov_gap}{numeric: Poverty gap.}

\item{A}{numeric: First regression parameter.}

\item{B}{numeric: Second regression parameter.}

\item{C}{numeric: Third regression parameter.}
}
\value{
numeric
}
\description{
Compute poverty severity for Lorenz Beta fit
}
\keyword{internal}
