% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_poverty_stats_lq}
\alias{gd_compute_poverty_stats_lq}
\title{Computes poverty stats from Lorenz Quadratic fit}
\usage{
gd_compute_poverty_stats_lq(mean, povline, ct)
}
\arguments{
\item{mean}{numeric: welfare mean}

\item{povline}{numeric: Poverty line}

\item{ct}{list: components from \code{get_components_lq}.}
}
\value{
list
}
\description{
Computes poverty stats from Lorenz Quadratic fit
}
