% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{gd_compute_quantile_lb}
\alias{gd_compute_quantile_lb}
\title{Compute quantiles from Lorenz Quandratic fit}
\usage{
gd_compute_quantile_lb(A, B, C, n_quantile = 10)
}
\arguments{
\item{A}{numeric vector: lorenz curve coefficient}

\item{B}{numeric vector: lorenz curve coefficient}

\item{C}{numeric vector: lorenz curve coefficient}
}
\value{
numeric
}
\description{
\code{gd_compute_quantile_lb()} computes quantiles from a lorenz beta fit.
}
