% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{gd_compute_quantile_lb}
\alias{gd_compute_quantile_lb}
\title{Compute quantiles from Lorenz Quandratic fit}
\usage{
gd_compute_quantile_lb(A, B, C, n_quantile = 10)
}
\arguments{
\item{A}{numeric: Lorenz curve coefficient.}

\item{B}{numeric: Lorenz curve coefficient.}

\item{C}{numeric: Lorenz curve coefficient.}

\item{n_quantile}{numeric: Number of quantiles to return.}
}
\value{
numeric
}
\description{
\code{gd_compute_quantile_lb()} computes quantiles from a Lorenz beta fit.
}
\keyword{internal}
