% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_quantile_lq}
\alias{gd_compute_quantile_lq}
\title{Compute quantiles from Lorenz Quandratic fit}
\usage{
gd_compute_quantile_lq(A, B, C, n_quantile = 10)
}
\arguments{
\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[3]}.}

\item{n_quantile}{numeric: Number of quantiles to return.}
}
\value{
numeric
}
\description{
\code{gd_compute_quantile_lq()} computes quantiles from a Lorenz Quadratic fit.
}
\keyword{internal}
