% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_quantile_lq}
\alias{gd_compute_quantile_lq}
\title{Compute quantiles from Lorenz Quandratic fit}
\usage{
gd_compute_quantile_lq(ct, n_quantile = 10)
}
\arguments{
\item{ct}{list: components from \code{get_components_lq}.}

\item{n_quantile}{numeric: number of quantiles. Default 10}
}
\value{
numeric
}
\description{
\code{gd_compute_quantile_lq()} computes quantiles from a Lorenz Quadratic fit.
}
