% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_quantile_lq}
\alias{gd_compute_quantile_lq}
\title{Compute quantiles from Lorenz Quandratic fit}
\usage{
gd_compute_quantile_lq(A, B, C, n_quantile = 10)
}
\arguments{
\item{A}{numeric vector: lorenz curve coefficient}

\item{B}{numeric vector: lorenz curve coefficient}

\item{C}{numeric vector: lorenz curve coefficient}

\item{n_quantile}{numeric: Number of quantiles to return}
}
\value{
numeric
}
\description{
\code{gd_compute_quantile_lq()} computes quantiles from a Lorenz Quadratic fit.
}
