% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{gd_compute_spl_lb}
\alias{gd_compute_spl_lb}
\title{Compute societal poverty line for group data from Lorenz beta fit}
\usage{
gd_compute_spl_lb(
  A = NULL,
  B = NULL,
  C = NULL,
  median = NULL,
  ppp_year = c(2017)
)
}
\arguments{
\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres()$coef[3]}.}

\item{median}{numeric median welfare}

\item{ppp_year}{numeric PPP year - 2017 (default) or 2011}
}
\value{
Societal poverty line
}
\description{
Compute societal poverty line for group data from Lorenz beta fit
}
