% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_watts_lq}
\alias{gd_compute_watts_lq}
\title{Computes Watts Index from Quadratic Lorenz fit}
\usage{
gd_compute_watts_lq(
  headcount,
  mu = lifecycle::deprecated(),
  mean = NULL,
  povline,
  dd = 0.01,
  A,
  B,
  C
)
}
\arguments{
\item{headcount}{numeric: Headcount index.}

\item{mu}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{mu} is no longer supported. Use
instead \code{mean}}

\item{mean}{numeric: mean of group data distribution}

\item{povline}{numeric: Poverty line.}

\item{dd}{numeric: \strong{TO BE DOCUMENTED}.}

\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[3]}.}
}
\value{
numeric
}
\description{
\code{gd_compute_watts_lq()} computes Watts Index from Quadratic Lorenz fit The
first distribution-sensitive poverty measure was proposed in 1968 by Watts It
is defined as the mean across the population of the proportionate poverty
gaps, as measured by the log of the ratio of the poverty line to income,
where the mean is formed over the whole population, counting the nonpoor as
having a zero poverty gap.
}
