% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_compute_watts_lq}
\alias{gd_compute_watts_lq}
\title{Computes Watts Index from Quadratic Lorenz fit}
\usage{
gd_compute_watts_lq(headcount, mu, povline, dd, A, B, C)
}
\arguments{
\item{headcount}{numeric: headcount index}

\item{mu}{numeric: \strong{TO BE DOCUMENTED}.}

\item{povline}{numeric: poverty line}

\item{dd}{numeric: \strong{TO BE DOCUMENTED}.}

\item{A}{numeric: Lorenz curve coefficient.}

\item{B}{numeric: Lorenz curve coefficient.}

\item{C}{numeric: Lorenz curve coefficient.}
}
\value{
numeric
}
\description{
\code{gd_compute_watts_lq()} computes Watts Index from Quadratic Lorenz fit
The first distribution-sensitive poverty measure was proposed in 1968 by Watts
It is defined as the mean across the population of the proportionate poverty
gaps, as measured by the log of the ratio of the poverty line to income,
where the mean is formed over the whole population, counting the nonpoor as
having a zero poverty gap.
}
\keyword{internal}
