% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_dist_stats.R
\name{gd_estimate_dist_stats_lq}
\alias{gd_estimate_dist_stats_lq}
\title{gd_estimate_dist_stats_lq
Estimates distributional stats from Quadratic Lorenz fit}
\usage{
gd_estimate_dist_stats_lq(mean, p0, A, B, C, key_values = key_values)
}
\arguments{
\item{mean}{numeric: Welfare mean.}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}

\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[3]}.}

\item{key_values}{named list: key values for lq fit, calculated using A, B, C
parameters using \link{gd_lq_key_values}}
}
\value{
list
}
\description{
gd_estimate_dist_stats_lq
Estimates distributional stats from Quadratic Lorenz fit
}
\keyword{internal}
