% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{gd_estimate_lb}
\alias{gd_estimate_lb}
\title{Estimates poverty and inequality stats from beta Lorenz fit}
\usage{
gd_estimate_lb(n_obs, mean, povline, p0, A, B, C)
}
\arguments{
\item{n_obs}{numeric: number of observations}

\item{mean}{numeric: Welfare mean}

\item{povline}{numeric: Poverty line}

\item{p0}{numeric: TO document}

\item{A}{numeric vector: Lorenz curve coefficient. Output of regres_lb()$coef\link{1}}

\item{B}{numeric vector: Lorenz curve coefficient. Output of regres_lb()$coef\link{2}}

\item{C}{numeric vector: Lorenz curve coefficient. Output of regres_lb()$coef\link{3}}
}
\value{
list
}
\description{
Estimates poverty and inequality stats from beta Lorenz fit
}
\examples{
estimate_lb(n_obs, mean, povline, p0, coefs)

}
