% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_estimate_lq}
\alias{gd_estimate_lq}
\title{Estimates poverty and inequality stats from Quadratic Lorenz fit}
\usage{
gd_estimate_lq(mean, povline, p0, A, B, C, key_values)
}
\arguments{
\item{mean}{numeric: Welfare mean.}

\item{povline}{numeric: Poverty line.}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}

\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[3]}.}

\item{key_values}{named list: key values for lq fit, calculated using A, B, C
parameters using \link{gd_lq_key_values}}
}
\value{
list
}
\description{
Estimates poverty and inequality stats from Quadratic Lorenz fit
}
\keyword{internal}
