% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{gd_estimate_lq}
\alias{gd_estimate_lq}
\title{Estimates poverty and inequality stats from Quadratic Lorenz fit}
\usage{
gd_estimate_lq(mean, povline, p0, ct)
}
\arguments{
\item{mean}{numeric: Welfare mean}

\item{povline}{numeric: Poverty line}

\item{p0}{numeric: TO document}

\item{A}{numeric vector: Lorenz curve coefficient. Output of regres_lq()$coef\link{1}}

\item{B}{numeric vector: Lorenz curve coefficient. Output of regres_lq()$coef\link{2}}

\item{C}{numeric vector: Lorenz curve coefficient. Output of regres_lq()$coef\link{3}}
}
\value{
list
}
\description{
Estimates poverty and inequality stats from Quadratic Lorenz fit
}
\examples{

}
