% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R, R/gd_estimate_lq.R
\name{gd_estimate_lq}
\alias{gd_estimate_lq}
\title{Estimates poverty and inequality stats from Quadratic Lorenz fit}
\usage{
gd_estimate_lq(n_obs, mean, povline, p0, A, B, C)

gd_estimate_lq(n_obs, mean, povline, p0, A, B, C)
}
\arguments{
\item{n_obs}{numeric: number of observations}

\item{mean}{numeric: Welfare mean}

\item{povline}{numeric: Poverty line}

\item{p0}{numeric: TO document}

\item{A}{numeric vector: Lorenz curve coefficient. Output of \code{regres_lq()$coef[1]}}

\item{B}{numeric vector: Lorenz curve coefficient. Output of \code{regres_lq()$coef[2]}}

\item{C}{numeric vector: Lorenz curve coefficient. Output of \code{regres_lq()$coef[3]}}
}
\value{
list

list
}
\description{
Estimates poverty and inequality stats from Quadratic Lorenz fit

Estimates poverty and inequality stats from Quadratic Lorenz fit
}
\keyword{internal}
