% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_select_lorenz.R
\name{gd_select_lorenz}
\alias{gd_select_lorenz}
\title{Select best Lorenz fit}
\usage{
gd_select_lorenz(lq, lb)
}
\arguments{
\item{lq}{list: Results from Lorenz Quadratic functional form. output of
\code{gd_compute_pip_stats_lq()}.}

\item{lb}{list: Results from Lorenz Beta functional form. output of
\code{gd_compute_pip_stats_lb()}.}
}
\value{
list
}
\description{
Select best Lorenz fit and adjust the returned statistics if needed.
}
\examples{
# Load example data
p <- system.file("tests/testdata/gd_ex2.RDS", package = "wbpip")
df <- readRDS(p)

# Estimate Quadratic Lorenz
lq <- wbpip:::gd_compute_pip_stats_lq(
    welfare = df$welfare,
    population = df$weight,
    requested_mean = 80,
    povline = 1.9,
    default_ppp = 1)

# Estimate Lorenz Beta
lb <- wbpip:::gd_compute_pip_stats_lb(
  welfare = df$welfare,
  population = df$weight,
  requested_mean = 50,
  povline = 1.9,
  default_ppp = 1)

# Select Lorenz
res <- wbpip:::gd_select_lorenz(lq, lb)

}
\keyword{internal}
