% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_select_lorenz.R
\name{gd_select_lorenz}
\alias{gd_select_lorenz}
\title{Select best Lorenz fit}
\usage{
gd_select_lorenz(lq, lb, n_quantile = 10)
}
\arguments{
\item{lq}{list: output of \code{gd_compute_pip_stats_lb()}}

\item{lb}{list: output of \code{gd_compute_pip_stats_lq()}}

\item{n_quantile}{numeric: Number of quantiles to be returned}
}
\value{
list
}
\description{
Select best Lorenz fit and adjust the returned statistics if needed
}
