% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_select_lorenz.R
\name{gd_select_lorenz}
\alias{gd_select_lorenz}
\title{Select best Lorenz fit}
\usage{
gd_select_lorenz(lq, lb)
}
\arguments{
\item{lq}{list: Results from Lorenz Quadratic functional form.
output of \code{gd_compute_pip_stats_lq()}}

\item{lb}{list: Results from Lorenz Beta functional form.
output of \code{gd_compute_pip_stats_lb()}}
}
\value{
list
}
\description{
Select best Lorenz fit and adjust the returned statistics if needed
}
