% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bins.R
\name{get_bins}
\alias{get_bins}
\title{get_bins}
\usage{
get_bins(
  .data,
  welfare,
  weight,
  type = "microdata",
  nbins = 100,
  output = "simple"
)
}
\arguments{
\item{.data}{Household survey dataframe with at least a welfare variable}

\item{welfare}{numeric: vector of welfare measures}

\item{weight}{numeric: vector of weights}

\item{type}{character: either 'microdata' or 'groupdata'. Defatult 'microdata'}

\item{nbins}{numeric: number of bins}

\item{output}{character: It as two variaties. \link{1} it could be a vector of
variables to retain after calculations (variables available are "welfare",
"weight", "cum_pop", "cum_prop_pop", and "bins"). \link{2} It could be a one of two
key words, "simple" or "full". \code{output = "simple"} is equivalent to
\code{output = "bins"} (which is the default). \code{output = "full"} if equivalent to
a vector with all the variables available,
\code{output = c("welfare", "weight", "cum_pop", "cum_prop_pop","bins")}}
}
\value{

}
\description{
Use function \code{md_compute_bins()} to calculate quantiles. In addition, this
function makes sure data is properly cleaned using \code{md_clean_data}
}
\examples{
df <- md_ABC_2000_income

bins <- get_bins(df, welfare, weight)
str(bins)

bins <- get_bins(df, welfare, weight, output = "full")
str(bins)
}
