% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gd.R
\name{get_gd_headcount_nv}
\alias{get_gd_headcount_nv}
\alias{get_gd_headcount}
\title{Poverty headcount for poverty line.}
\usage{
get_gd_headcount_nv(
  welfare = NULL,
  population = NULL,
  params = NULL,
  mean = 1,
  times_mean = 1,
  povline = mean * times_mean,
  complete = FALSE,
  lorenz = NULL
)

get_gd_headcount(
  welfare = NULL,
  population = NULL,
  params = NULL,
  mean = 1,
  times_mean = 1,
  povline = mean * times_mean,
  complete = FALSE,
  lorenz = NULL
)
}
\arguments{
\item{welfare}{numeric: cumulative sahre of welfare (income/consumption)}

\item{population}{numeric: cumulative share of the population}

\item{params}{list of parameters from \code{get_gd_lorenz_validity()}}

\item{mean}{numeric: welfare mean of distribution.}

\item{povline}{numeric: value of poverty line. Default is the \code{mean} value}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}

\item{lorenz}{character or NULL. Lorenz curve selected. It could be "lq" for
Lorenz Quadratic or "lb" for Lorenz Beta}
}
\value{
list with headcount and selected Lorenz
}
\description{
There are two version. \code{get_gd_headcount_nv} which is not vectorized, and
\code{get_gd_headcount}, which is vectorize in the \code{povline} argument. There is no
need to use \code{get_gd_headcount_nv} unless it is for debuging purposes given
that it is the original function. \code{get_gd_headcount} is difficult to debug
}
