% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gd.R
\name{get_gd_pov_gap_nv}
\alias{get_gd_pov_gap_nv}
\alias{get_gd_pov_gap}
\title{Group Data Poverty Gap}
\usage{
get_gd_pov_gap_nv(
  welfare = NULL,
  population = NULL,
  params = NULL,
  mean = 1,
  times_mean = 1,
  povline = mean * times_mean,
  complete = FALSE,
  lorenz = NULL
)

get_gd_pov_gap(
  welfare = NULL,
  population = NULL,
  params = NULL,
  mean = 1,
  times_mean = 1,
  povline = mean * times_mean,
  complete = FALSE,
  lorenz = NULL,
  format = c("dt", "list", "atomic")
)
}
\arguments{
\item{welfare}{numeric: cumulative sahre of welfare (income/consumption)}

\item{population}{numeric: cumulative share of the population}

\item{params}{list of parameters from \code{get_gd_lorenz_validity()}}

\item{mean}{numeric: welfare mean of distribution.}

\item{povline}{numeric: value of poverty line. Default is the \code{mean} value}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}

\item{lorenz}{character or NULL. Lorenz curve selected. It could be "lq" for
Lorenz Quadratic or "lb" for Lorenz Beta}

\item{format}{character: either "dt" for data.table, "list" or "atomic" for a
single numeric vector, whose names are corresponding selected Lorenz for
each value.  Default is "dt"}
}
\value{
data.table, list or atomic vector

data.table with poverty headcounts and selected Lorenz
}
\description{
Group Data Poverty Gap
}
\examples{

# Return data.table
get_gd_pov_gap(
welfare = grouped_data_ex2$welfare,
population = grouped_data_ex2$weight,
povline = c(.5, 1, 2, 3))

# Return list
get_gd_pov_gap(
welfare = grouped_data_ex2$welfare,
population = grouped_data_ex2$weight,
povline = c(.5, 1, 2, 3),
format = "list")

# Return list complete
get_gd_pov_gap(
welfare = grouped_data_ex2$welfare,
population = grouped_data_ex2$weight,
povline = c(.5, 1, 2, 3),
format = "list",
complete = TRUE)

# Return data.table
get_gd_pov_gap(
welfare = grouped_data_ex2$welfare,
population = grouped_data_ex2$weight,
povline = c(.5, 1, 2, 3),
format = "atomic")
}
