% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gd.R
\name{get_gd_quantiles}
\alias{get_gd_quantiles}
\title{calculate quantiles}
\usage{
get_gd_quantiles(
  welfare = NULL,
  population = NULL,
  params = NULL,
  mean = 1,
  times_mean = 1,
  povline = mean * times_mean,
  complete = FALSE,
  lorenz = NULL,
  n = 10
)
}
\arguments{
\item{welfare}{numeric: cumulative sahre of welfare (income/consumption)}

\item{population}{numeric: cumulative share of the population}

\item{params}{list of parameters from \code{get_gd_lorenz_params()}}

\item{mean}{numeric: welfare mean of distribution.}

\item{povline}{numeric: value of poverty line. Default is the \code{mean} value}

\item{complete}{logical: If TRUE, returns a list a cumulative returns from
previously used \code{get_gd} functions. Default is \code{FALSE}}

\item{lorenz}{character or NULL. Lorenz curve selected. It could be "lq" for
Lorenz Quadratic or "lb" for Lorenz Beta}

\item{n}{numeric: Number of quantiles}
}
\value{
list with vector of quantiles
}
\description{
calculate quantiles
}
