% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gini.R
\name{get_gini}
\alias{get_gini}
\title{Title}
\usage{
get_gini(.data, welfare, weight, type = "microdata")
}
\arguments{
\item{.data}{Household survey dataframe with at least a welfare variable}

\item{welfare}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights, optional.}

\item{type}{character: either 'microdat' or 'groupdata'. Defatult 'microdata'}
}
\value{

}
\description{
Title
}
\examples{
get_gini(dt, welfare, weight)

setnames(md_ABC_2000_income, c("welfare", "weight"), c("wf", "wt") )

md_ABC_2000_income \%>\%
  group_by(area) \%>\%
  get_gini(wf, wt)

md_ABC_2000_income \%>\%
  get_gini(wf, wt)
}
