% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gini.R
\name{get_gini}
\alias{get_gini}
\title{Title}
\usage{
get_gini(.data, welfare, weight, type = "microdata")
}
\arguments{
\item{.data}{Household survey dataframe with at least a welfare variable}

\item{welfare}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights, optional.}

\item{type}{character: either 'microdata' or 'groupdata'. Defatult 'microdata'}
}
\value{

}
\description{
Title
}
