% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lorenz.R
\name{get_lorenz}
\alias{get_lorenz}
\title{Lorenz curve}
\usage{
get_lorenz(
  .data,
  welfare,
  weight,
  distribution_type = c("micro", "group", "aggregate", "imputed"),
  nbins = NULL
)
}
\arguments{
\item{.data}{Household survey data frame with at least a welfare variable.}

\item{welfare}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights.}

\item{distribution_type}{character: Type of distribution, either micro,
group, aggregate or imputed.}

\item{nbins}{numeric: number of points on the Lorenz curve.}
}
\value{
data.frame
}
\description{
Calculate points on the Lorenz curve. \strong{TO BE DOCUMENTED}.
}
\examples{
data("md_ABC_2000_income")
df     <- md_ABC_2000_income
lorenz <- get_lorenz(df, welfare, weight)
str(lorenz)

}
