% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lorenz.R
\name{get_lorenz}
\alias{get_lorenz}
\title{get_lorenz}
\usage{
get_lorenz(.data, welfare, weight, type = "microdata", nbins = NULL)
}
\arguments{
\item{.data}{Household survey dataframe with at least a welfare variable}

\item{welfare}{numeric: vector of welfare measures}

\item{weight}{numeric: vector of weights}

\item{type}{character: either 'microdata' or 'groupdata'. Defatult 'microdata'}

\item{nbins}{numeric: number of points on the Lorenz curve}
}
\value{
data.frame
}
\description{
Use function \code{md_compute_lorenz()} to calculate quantiles. In addition, this
function makes sure data is properly cleaned using \code{md_clean_data}
}
