% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lorenz_alt.R
\name{lorenz_alt}
\alias{lorenz_alt}
\title{lorenz}
\usage{
lorenz_alt(welfare, weight = NULL, nbins = NULL, na.rm = FALSE)
}
\arguments{
\item{welfare}{numeric: vector of welfare measures}

\item{weight}{numeric: vector of weights}

\item{nbins}{numeric: number of points on the Lorenz curve}

\item{na.rm}{logical: Set to TRUE to remove missing}
}
\value{
data.frame
}
\description{
Given a vector of weights and welfare, this functions computes the
Lorenz curve and selveral other related indicators.
}
\examples{
data("md_ABC_2010_income")
df <- md_ABC_2010_income
lz <- lorenz_alt(df$welfare, df$weight)[]
}
