% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_clean_data.R
\name{md_clean_data}
\alias{md_clean_data}
\title{Check microdata to be used in PIP methods}
\usage{
md_clean_data(dt, ...)
}
\arguments{
\item{dt}{Data frame.}

\item{...}{list of arguments that correspond to specific variable names.
Arguments available are in \code{getOption("wbpip.agrs_to_check")}.
For instance, welfare = "income", weight = "peso'.}
}
\value{
list of elements whose main object is a dataframe (in data.table format)
with the necessary transformations to be included in PIP methods. Data is available in
element $data. The other elements provide the number of observations that were modified
depending on test performed. The name of elements are in the form p_s, where p (or prefix)
refers to the test, whereas s (for suffix) refers to the name of the variable evaluated.
Prefix are:

\emph{nna:}  Number of NA in variable
\emph{nng:}  Number of negative values
\emph{ina:}  Index of obs with NA in variable
\emph{ing:}  Index of obs with negative values
}
\description{
Check microdata to be used in PIP methods
}
