% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_dist_stats.R
\name{md_compute_dist_stats}
\alias{md_compute_dist_stats}
\title{Compute distributional statistics}
\usage{
md_compute_dist_stats(
  welfare,
  weight,
  mean = NULL,
  nbins = 10,
  lorenz = NULL,
  n_quantile = 10
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{mean}{numeric: A value with the mean. Optional.}

\item{nbins}{numeric: number of points on the Lorenz curve. Optional. Only
used if \code{lorenz} is NULL. Default is 10.}

\item{lorenz}{data.frame: A table with points on the Lorenz curve. Output of
\code{md_compute_lorenz()}. Optional.}

\item{n_quantile}{numeric: Number of quantiles for which share of total
income is desired. It can't be larger that the total number of percentiles
in the Lorenz curve provided by the user. Default is 10.}
}
\value{
data.frame
}
\description{
Given a vector of weights and welfare, this functions computes the various
distributional statistics.
}
\examples{
wbpip:::md_compute_dist_stats(welfare = 1:2000, weight = rep(1, 2000))
}
\keyword{internal}
