% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_dist_stats.R
\name{md_compute_dist_stats}
\alias{md_compute_dist_stats}
\title{Compute distributionla statistics}
\usage{
md_compute_dist_stats(welfare, weight = NULL, nbins = NULL, n_quantile = 10)
}
\arguments{
\item{welfare}{numeric: vector of welfare measures}

\item{weight}{numeric: vector of weights}

\item{nbins}{numeric: number of points on the Lorenz curve}

\item{n_quantile}{numeric: Number of quantiles for which share of total income
is desired. It can't be larger that the total number of percentiles in the
Lorenz curve provided by the user.  default is 10.}
}
\value{
data.frame

#' @examples#'
data("md_ABC_2000_income")
md_compute_dist_stats(md_ABC_2000_income$welfare, md_ABC_2000_income$weight)
}
\description{
Given a vector of weights and welfare, this functions computes the
various distributional statistics
}
