% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_poverty_stats.R
\name{md_compute_fgt}
\alias{md_compute_fgt}
\title{Estimate FGT measures from microdata}
\usage{
md_compute_fgt(welfare, weight, pl, alpha = 0)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{pl}{numeric: poverty line}

\item{alpha}{numeric: either 0, 1 or 2.}
}
\value{
numeric vector of length 1
}
\description{
Estimate FGT measures from microdata
}
\examples{
md_compute_fgt(
  welfare = 1:2000,
  weight = rep(1, 2000),
  pl = 10,
  alpha = 0
)
}
