% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_gini.R
\name{md_compute_gini}
\alias{md_compute_gini}
\title{Gini coefficient}
\usage{
md_compute_gini(measure, weight)
}
\arguments{
\item{measure}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights.}
}
\value{
numeric
}
\description{
Compute the Gini coefficient.
}
\details{
Given a vector of income or consumption values and their respective weights
\code{md_compute_gini()} computes the Gini coefficient for the distribution.

Note that the input parameters cannot contain missing or negative values and
that \code{measure} must be sorted in increasing order.
}
\examples{
# Simple example
md_compute_gini(measure = 1:100, weight = rep(1, 100))

}
