% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_poverty_stats.R,
%   R/md_compute_poverty_stats_replacement.R
\name{md_compute_headcount}
\alias{md_compute_headcount}
\title{Compute headcount from microdata}
\usage{
md_compute_headcount(
  welfare = NULL,
  weight = NULL,
  povline = NULL,
  weight_pov = NULL,
  weight_total = NULL,
  verbose = FALSE
)

md_compute_headcount(
  welfare = NULL,
  weight = NULL,
  povline = NULL,
  weight_pov = NULL,
  weight_total = NULL,
  verbose = FALSE
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{weight}{numeric: A vector of population weights, optional, a vector of 1s if not specified.}

\item{povline}{numeric: Poverty line in international dollars}

\item{weight_pov}{numeric: A vector of population weights for the population
in poverty. Default is NULL, primary purpose is internal}

\item{weight_total}{numeric: sum of population weights}

\item{verbose}{logical: display messages. Default is FALSE}
}
\value{
numeric

numeric
}
\description{
Compute headcount from microdata

Compute headcount from microdata
}
