% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_lorenz.R
\name{md_compute_lorenz}
\alias{md_compute_lorenz}
\title{Lorenz curve}
\usage{
md_compute_lorenz(
  welfare,
  weight = rep(1, length(welfare)),
  nbins = if (length(welfare) > 1000) 100 else 20,
  type = 7L
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights. Default is a vector of ones,
\code{rep(1, length(welfare))}.}

\item{nbins}{numeric: number of points on the Lorenz curve. if \code{NULL} the
returning  Lorenz curve would be the  length of the original welfare vector
minus the number of \code{NAs} of different observations in  \code{welfare} and
\code{weight} plus 1 to account for the (0,0) intercept. Default is \code{100} for
\code{length(welfare) > 1000} and \code{20} otherwise.}

\item{type}{numeric: integer. Quantile types 5-9. See \link{quantile}. The default
method is type 7.}
}
\value{
data.frame
}
\description{
Compute the Lorenz curve for microdata.
}
\details{
Given a vector of weights and welfare, this functions computes the Lorenz
curve.
}
\examples{
wbpip:::md_compute_lorenz(welfare = 1:2000, weight = rep(1, 2000))
}
