% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_lorenz.R
\name{md_compute_lorenz}
\alias{md_compute_lorenz}
\title{md_compute_lorenz
Given a vector of weights and welfare, this functions computes the
Lorenz curve}
\usage{
md_compute_lorenz(welfare, weight = NULL, nbins = NULL)
}
\arguments{
\item{welfare}{numeric: vector of welfare measures}

\item{weight}{numeric: vector of weights}

\item{nbins}{numeric: number of points on the Lorenz curve}
}
\value{
data.frame

#' @examples
md_compute_lorenz(welfare = 1:2000, weight = 1:2000)

data("md_ABC_2000_income")
md_compute_lorenz(md_ABC_2000_income$welfare, md_ABC_2000_income$weight)
}
\description{
md_compute_lorenz
Given a vector of weights and welfare, this functions computes the
Lorenz curve
}
