% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_quantiles.R
\name{md_compute_median}
\alias{md_compute_median}
\title{Compute median for microdata}
\usage{
md_compute_median(welfare, weight)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights. Default is a vector of ones}
}
\value{
numeric
}
\description{
Compute median for microdata
}
\examples{
md_compute_median(welfare = 1:2000, weight = rep(1, 2000))
}
\keyword{internal}
