% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_median.R
\name{md_compute_median}
\alias{md_compute_median}
\title{compute median for microdata}
\usage{
md_compute_median(welfare, weight)
}
\arguments{
\item{welfare}{welfare vector}

\item{weight}{population weights vector}
}
\value{
scalar
}
\description{
compute median for microdata
}
\keyword{internal}
