% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_mld.R
\name{md_compute_mld}
\alias{md_compute_mld}
\title{md_compute_mld
Given a vector of weights and welfare, this functions computes the
Mean Log Deviation (MLD)}
\usage{
md_compute_mld(welfare, weight = NULL)
}
\arguments{
\item{welfare}{numeric: vector of welfare measures}

\item{weight}{numeric: vector of weights}
}
\value{
numeric
}
\description{
md_compute_mld
Given a vector of weights and welfare, this functions computes the
Mean Log Deviation (MLD)
}
\examples{
md_compute_mld(welfare = 1:2000, weight = 1:2000)

data("md_ABC_2000_income")
md_compute_mld(md_ABC_2000_income$welfare, md_ABC_2000_income$weight)
}
