% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_mld.R
\name{md_compute_mld}
\alias{md_compute_mld}
\title{Mean Log Deviation}
\usage{
md_compute_mld(welfare, weight = NULL)
}
\arguments{
\item{welfare}{numeric: vector of welfare measures}

\item{weight}{numeric: vector of weights}
}
\value{
numeric
}
\description{
Given a vector of weights and welfare, this functions computes the
Mean Log Deviation (MLD).
}
\examples{
wbpip:::md_compute_mld(welfare = 1:2000)

}
\keyword{internal}
