% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_pip_stats.R
\name{md_compute_pip_stats}
\alias{md_compute_pip_stats}
\title{Compute default PIP statistics}
\usage{
md_compute_pip_stats(
  welfare,
  povline,
  population,
  requested_mean = NULL,
  popshare = NULL,
  default_ppp = 1,
  ppp = NULL,
  ppp_year = 2017,
  cons_floor = 0.5
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{povline}{numeric: Poverty line in international dollars}

\item{population}{numeric: A vector of population weights, optional, a vector
of 1s if not specified.}

\item{requested_mean}{numeric: Welfare mean in international dollars}

\item{popshare}{numeric: Share of population for which the corresponding
quantile is desired. Default .5 (i.e., weighted median)}

\item{default_ppp}{numeric: Default purchasing power parity}

\item{ppp}{numeric: PPP requested by user}

\item{ppp_year}{numeric: PPP year - 2017 (default) or 2011}

\item{cons_floor}{numeric: gives the consumption floor. Default is $0.5 per day, if NULL then no floor.}
}
\value{
list
}
\description{
Compute poverty and distributional statistics for microdata.
}
\keyword{internal}
