% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_pip_stats.R
\name{md_compute_pip_stats}
\alias{md_compute_pip_stats}
\title{Compute default PIP statistics from microdata}
\usage{
md_compute_pip_stats(
  welfare,
  povline,
  weight = NULL,
  ppp = NULL,
  default_ppp = 1,
  requested_mean = NULL,
  popshare = NULL
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{povline}{numeric: Monthly poverty line in international dollars}

\item{weight}{numeric: A vector of weights, optional, a vector of 1s if not specified.}

\item{ppp}{numeric: PPP requested by user}

\item{default_ppp}{numeric: Default purchasing power parity}

\item{requested_mean}{numeric: Welfare mean in international dollars}

\item{popshare}{numeric: Share of population for which the corresponding
quantile is desired. Default .5 (i.e., weighted median)}
}
\value{
list
}
\description{
Compute poverty and distributional statistics from microdata records
}
