% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_polarization.R
\name{md_compute_polarization}
\alias{md_compute_polarization}
\title{Wolfson polarization index}
\usage{
md_compute_polarization(welfare, weight, gini, mean, median, cons_floor = 0.5)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{gini}{numeric: Gini. Output of \code{\link[=md_compute_gini]{md_compute_gini()}}.}

\item{mean}{numeric: A value with the mean. Optional.}

\item{median}{numeric: Median. Output of \code{\link[=md_compute_quantiles]{md_compute_quantiles()}}.}
}
\value{
numeric
}
\description{
Compute the Wolfson polarization index for microdata.
}
\examples{
wbpip:::md_compute_polarization(
  welfare = 1:2000,
  weight = rep(1, 2000),
  gini = 0.4,
  mean = 950,
  median = 1000
)
}
\references{
Ravallion, M., S. Chen. 1996.
"\href{http://documents1.worldbank.org/curated/en/202781468739531561/pdf/multi-page.pdf}{What Can New Survey Data Tell Us about Recent Changes in Distribution and Poverty?}".
Policy Research Working Paper 1694.
World Bank, Washington, DC.
}
\keyword{internal}
