% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_polarization.R
\name{md_compute_polarization}
\alias{md_compute_polarization}
\title{Wolfson polarization index}
\usage{
md_compute_polarization(welfare, weight = NULL, gini, weighted_median)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights, optional.}

\item{gini}{numeric: Gini. Output of \code{\link[=md_compute_gini]{md_compute_gini()}}.}

\item{weighted_median}{numeric: Median. Output of \code{\link[=md_compute_quantiles]{md_compute_quantiles()}}.}
}
\description{
Compute the Wolfson polarization index for microdata.
}
\examples{
# Simple example
md_compute_polarization(welfare = 1:10)

# Include weights
md_compute_polarization(welfare = 1:10, weight = 1:10)

# Microdata example
data("md_ABC_2000_income")
md_compute_polarization(
  md_ABC_2000_income$welfare,
  md_ABC_2000_income$weight)

}
\references{
Ravallion, M., S. Chen. 1996.
"\href{http://documents1.worldbank.org/curated/en/202781468739531561/pdf/multi-page.pdf}{What Can New Survey Data Tell Us about Recent Changes in Distribution and Poverty?}".
Policy Research Working Paper 1694.
World Bank, Washington, DC.
}
\keyword{internal}
