% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_poverty_stats.R
\name{md_compute_poverty_stats}
\alias{md_compute_poverty_stats}
\title{Compute Poverty Statistics}
\usage{
md_compute_poverty_stats(welfare, povline, weight = NULL)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values}

\item{povline}{numeric: A poverty line/threshold}

\item{weight}{numeric: A vector of weights, optional, a vector of 1s if not specified.}
}
\value{
list containing poverty headcount, poverty gap, poverty severity and the watts index
}
\description{
Compute the poverty headcount, poverty gap, poverty severity and
the watts index for a micro data set.
}
\details{
Given a vector of consumption or income values and their respective weights
\code{md_compute_poverty_stats()} computes poverty headcount, poverty gap,
poverty severity and the watts index.

Negative values are dropped from computations.
}
\examples{
# simple example (no weights)
md_compute_poverty_stats(welfare = 1:100, povline = 10)

# with weights
md_compute_poverty_stats(welfare = 1:100, povline = 10, weight = 1:100)

# using a micro data set
data("md_ABC_2000_income")
md_compute_poverty_stats(md_ABC_2000_income$welfare, 2500000, md_ABC_2000_income$weight)


}
