% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_povline_lcu.R
\name{md_compute_povline_lcu}
\alias{md_compute_povline_lcu}
\title{Compute poverty line in Local Currency Unit (LCU)}
\usage{
md_compute_povline_lcu(
  welfare,
  povline,
  weight = NULL,
  popshare,
  requested_mean,
  data_mean
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values.}

\item{povline}{numeric: Poverty line in internationl dollars (PPP)}

\item{weight}{numeric: A vector of weights, optional.}

\item{popshare}{umeric: Share of population for which the corresponding
quantile is desired. Default .5 (i.e., weighted median)}

\item{requested_mean}{numeric: requested mean (adjusted with PPPs)}

\item{data_mean}{numeric: Data mean}
}
\value{
list
}
\description{
Compute poverty line in Local Currency Unit (LCU)
}
