% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_quantiles.R
\name{md_compute_quantiles}
\alias{md_compute_quantiles}
\title{Compute quantiles for microdata with lorenz function}
\usage{
md_compute_quantiles(
  welfare = NULL,
  weight = rep(1, length(welfare)),
  n_quantile = 10,
  lorenz = NULL
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights. Default is a vector of ones,}

\item{lorenz}{numeric: Output from \code{md_compute_lorenz}}

\item{n_quantiles}{numeric: Number of quantiles for which share of total income
is desired. Default is 10.}
}
\value{
list
}
\description{
Compute quantiles for microdata with lorenz function
}
\examples{
md_compute_quantiles(welfare = 1:2000, weight = rep(1, 2000))
}
\keyword{internal}
