% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_quantiles.R
\name{md_compute_quantiles_c}
\alias{md_compute_quantiles_c}
\title{Compute quantiles for microdata with collapse}
\usage{
md_compute_quantiles_c(
  welfare,
  weight = rep(1, length(welfare)),
  n_quantile = 10
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights. Default is a vector of ones,}

\item{n_quantiles}{numeric: Number of quantiles for which share of total income
is desired. Default is 10.}
}
\value{
list
}
\description{
Compute quantiles for microdata with collapse
}
\keyword{internal}
