% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_quantiles.R
\name{md_compute_quantiles_share}
\alias{md_compute_quantiles_share}
\title{Compute quantiles share}
\usage{
md_compute_quantiles_share(welfare, weight, n_quantile = 10)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights. Default is a vector of ones}

\item{n_quantile}{numeric: Number of quantiles for which share of total income
is desired. It can't be larger that the total number of percentiles in the
Lorenz curve provided by the user.  default is 10.}
}
\value{
list
}
\description{
Compute quantiles for microdata.
}
\details{
Calculate share of total welfare in each quantile from Lorenz curve and
the its corresponding percentile. That is, it requires a vector with the
cumulative share of the population (\code{lweight}), a vector with cumulative
share of welfare (\code{lwelfare}), and a vector the corresponding monetary value
of each percentile (\code{percentile}).
}
\examples{
md_compute_quantiles_share(welfare = 1:2000, weight = rep(1, 2000))

}
\keyword{internal}
