% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_compute_spl.R
\name{md_compute_spl}
\alias{md_compute_spl}
\title{Compute SPL}
\usage{
md_compute_spl(welfare, weight, threshold_rate = 0.5, ppp = 1, cpi = 1)
}
\arguments{
\item{welfare}{numeric: A vector of consumption or income values}

\item{weight}{numeric: A vector of weights corresponding for each welfare value}

\item{threshold_rate}{numeric : A value between 0 and 1 for the proportion of
the median consumption/income for SPL calculation}

\item{ppp}{numeric: PPP values inputed by the user (set to 1 if welfare values are already in dollars)}

\item{cpi}{numeric: CPI values inputed by the user (set to 1 if welfare values year is the reference year)}
}
\value{
numeric
}
\description{
Compute the societal poverty line based on the median daily consumption in US dollars
}
\keyword{internal}
