% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_infer_poverty_line.R
\name{md_infer_poverty_line}
\alias{md_infer_poverty_line}
\title{md_infer_poverty_line}
\usage{
md_infer_poverty_line(
  welfare,
  weight = NULL,
  popshare = 0.5,
  na.rm = FALSE,
  include = FALSE
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights, optional.}

\item{popshare}{numeric: Share of population for which the corresponding
quantile is desired. Default .5 (i.e., weighted median)}

\item{na.rm}{logical: Set to TRUE to remove missing from computations}

\item{include}{logical:}
}
\value{
numeric
}
\description{
calculate percentiles corresponding to the specified  Shared of population
(percentages)
}
\examples{
df <- md_ABC_2000_income
# median
md_infer_poverty_line(df$welfare, df$weight)
md_infer_poverty_line(df$welfare, df$weight, popshare = .2)
md_infer_poverty_line(df$welfare, df$weight, popshare = .6)

}
