% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_quantile_functions.R
\name{md_quantile_values}
\alias{md_quantile_values}
\title{Get quantile at specified shared of population - micro data}
\usage{
md_quantile_values(
  welfare = NULL,
  weight = rep(1, length = length(welfare)),
  n = 10,
  format = c("dt", "list", "atomic")
)
}
\arguments{
\item{welfare}{welfare vector}

\item{weight}{population weight vector. Default is 1}

\item{n}{numeric: number of equi-spaced quantiles}

\item{format}{character: "dt", "list", "atomic", giving the format of the
output}
}
\value{
quantiles: see \code{format}
}
\description{
\code{md_quantile_values} returns the quantile (i.e., monetary value) that corresponds
to the share of the population that lives below that threshold.
}
\details{
The quantiles are calculated using the \code{welfare} outcome of the function
\code{md_compute_lorenz}.
}
\examples{
md_quantile_values(
  welfare = md_GHI_2000_consumption$welfare,
  weight  = md_GHI_2000_consumption$weight,
  n       = 5
)
}
