% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_quantile_functions.R
\name{md_quantile_values}
\alias{md_quantile_values}
\title{Get quantile at specified shared of population - micro data}
\usage{
md_quantile_values(
  welfare = NULL,
  weight = rep(1, length = length(welfare)),
  n = NULL,
  popshare = ifelse(is.null(n), seq(from = 1/10, to = 1, by = 1/10), seq(from = 1/n, to =
    1, by = 1/n)),
  format = c("dt", "list", "atomic")
)
}
\arguments{
\item{welfare}{welfare vector}

\item{weight}{population weight vector. Default is 1}

\item{n}{numeric: number of equi-spaced quantiles}

\item{popshare}{numeric atomic vector: the quantiles to return. Will only be
used if \code{n = NULL}}

\item{format}{character: "dt", "list", "atomic", giving the format of the
output}
}
\value{
quantiles: see \code{format}
}
\description{
\code{md_quantile_values} returns the quantile (i.e., monetary value) that corresponds
to share of the population that lives below that threshold.
}
\details{
This is basically the inverse of estimating the poverty rate (headcount or
population share) below the poverty line. In this case, you provide the
headcount and \code{md_quantile_values} returns the "poverty line".

The quantiles are calculated as function of the mean of the distribution
times an \code{x} factor. Basically, the quantile is \code{x} times the mean. By
default, the mean is equal to 1, which implies that, if no mean value if
provided, the return value is equal to \code{x}.
}
\examples{
md_quantile_values(
  welfare = md_GHI_2000_consumption$welfare
  weight  = md_GHI_2000_consumption$weight,
  n       = 5
)
}
