% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_quantile_functions.R
\name{md_quantile_welfare_share}
\alias{md_quantile_welfare_share}
\title{Quantile welfare share}
\usage{
md_quantile_welfare_share(
  welfare = NULL,
  weight = rep(1, length = length(welfare)),
  n = NULL,
  format = c("dt", "list", "atomic")
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values.}

\item{weight}{numeric: A vector of weights. Default is a vector of ones,
\code{rep(1, length(welfare))}.}
}
\value{
list with vector of share of welfare by quantiles
}
\description{
\code{md_quantile_welfare_share} returns the share of welfare held by a
particular quantile. Notice that \code{md_welfare_share_at} get the share of
welfare held by a particular share of the population, which is in a sense
the cumulative share. Instead, \code{md_quantile_welfare_share} returns
the proportion of welfare that only the specified quantile holds.
}
\examples{
md_quantile_welfare_share(welfare = md_GHI_2000_consumption$welfare,
                             weight = md_GHI_2000_consumption$weight)
}
