% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_quantile_functions.R
\name{md_quantile_welfare_share}
\alias{md_quantile_welfare_share}
\title{Quantile welfare share}
\usage{
md_quantile_welfare_share(
  welfare = NULL,
  weight = rep(1, length = length(welfare)),
  n = NULL,
  popshare = ifelse(is.null(n), seq(from = 1/10, to = 1, by = 1/10), seq(from = 1/n, to =
    1, by = 1/n)),
  format = c("dt", "list", "atomic")
)
}
\arguments{
\item{welfare}{welfare vector}

\item{weight}{population weight vector. Default is 1}

\item{n}{numeric: number of equi-spaced quantiles}

\item{popshare}{numeric atomic vector: the quantiles to return. Will only be
used if \code{n = NULL}}

\item{format}{character: "dt", "list", "atomic", giving the format of the
output}
}
\value{
list with vector of share of welfare by quantiles
}
\description{
\code{md_quantile_welfare_share} returns the share of welfare held by a
particular quantile. Notice that \code{md_welfare_share_at} get the share of
welfare held by a particular share of the population, which is in a sense
the cumulative share. Instead, \code{md_quantile_welfare_share} returns
the proportion of welfare that only the specified quantile holds.
}
\examples{
md_quantile_welfare_share(welfare = md_GHI_2000_consumption$welfare,
                             weight = md_GHI_2000_consumption$weight)
}
