% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_quantile_functions.R
\name{md_welfare_share_at}
\alias{md_welfare_share_at}
\title{Welfare share by quantile in micro data}
\usage{
md_welfare_share_at(
  welfare = NULL,
  weight = rep(1, length = length(welfare)),
  n = NULL,
  popshare = ifelse(is.null(n), seq(from = 1/10, to = 1, by = 1/10), seq(from = 1/n, to =
    1, by = 1/n)),
  format = c("dt", "list", "atomic")
)
}
\arguments{
\item{welfare}{welfare vector}

\item{weight}{population weight vector. Default is 1}

\item{n}{numeric: number of equi-spaced quantiles}

\item{popshare}{numeric atomic vector: the quantiles to return. Will only be
used if \code{n = NULL}}

\item{format}{character: "dt", "list", "atomic", giving the format of the
output}
}
\value{
list with vector of share of welfare by quantiles
}
\description{
\code{md_welfare_share_at} returns the share of welfare held by the specified
share of the population in the parameter \code{popshare}. Alternatively, you can
select the number of quantiles (10 be default), to estimate the corresponding
share of welfare in each.
}
\examples{
md_welfare_share_at(welfare = md_GHI_2000_consumption$welfare,
                    weight = md_GHI_2000_consumption$weight)
}
