% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_md_compute_pip_stats.R
\name{prod_md_compute_pip_stats}
\alias{prod_md_compute_pip_stats}
\title{Compute default PIP statistics
Version used in production. Does not re-compute all distributional stats.
Distributional stats are pre-computed as they are not sensitive to the poverty line.}
\usage{
prod_md_compute_pip_stats(
  welfare,
  povline,
  population = NULL,
  requested_mean = NULL,
  popshare = NULL,
  default_ppp = 1,
  ppp = NULL
)
}
\arguments{
\item{welfare}{numeric: A vector of income or consumption values.}

\item{povline}{numeric: Monthly poverty line in international dollars.}

\item{population}{numeric: A vector of population weights, optional, a vector
of 1s if not specified.}

\item{requested_mean}{numeric: Welfare mean in international dollars}

\item{popshare}{numeric: Share of population for which the corresponding
quantile is desired. Default .5 (i.e., weighted median).}

\item{default_ppp}{numeric: Default purchasing power parity.}

\item{ppp}{numeric: PPP requested by user.}
}
\value{
list
}
\description{
Compute poverty and distributional statistics for microdata.
}
\keyword{internal}
