% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_utils.R
\name{regres}
\alias{regres}
\title{Perform linear regression on Lorenz formatted input}
\usage{
regres(df, is_lq = TRUE)
}
\arguments{
\item{df}{data.frame: Output of \code{create_functional_form_lq()} or
\code{create_functional_form_lb()}.}

\item{is_lq}{logical: TRUE if Lorenz Quadratic, FALSE if Beta Lorenz.}
}
\value{
list
}
\description{
\code{regres()} performs linear regression on lorenz formatted input (Beta or
Quadratic). There is no intercept in the regression. The coefficients of
regressions are estimated by ordinary least squares.
}
\keyword{internal}
