% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_create_synth_vector.R
\name{sd_create_synth_vector}
\alias{sd_create_synth_vector}
\title{Create synthetic vector based on Lorenz  parameters of group data}
\usage{
sd_create_synth_vector(
  welfare,
  population,
  mean,
  pop = NULL,
  p0 = 0.5,
  nobs = 100000,
  selected_model = NULL,
  verbose = FALSE
)
}
\arguments{
\item{welfare}{numeric: Cumulative proportion of welfare held by that
proportion of the population (Lorenz Curve).}

\item{population}{numeric: Cumulative proportion of population.}

\item{mean}{numeric: Welfare mean.}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}

\item{nobs}{numeric: Number of observations to be used in synthetic vector}

\item{selected_model}{character or NULL: force to use one model, either
quadratic (q) or Beta (b). If NULL (default), the methodological protocol
will select the model. This argument is only useful for testing purposes.}
}
\value{
list
}
\description{
Compute distributional statistics for grouped data by selecting the best
functional fit for the Lorenz curve (either beta or quadratic).
}
\examples{
# Create synthetic dataset
res <- wbpip:::sd_create_synth_vector(
  welfare = grouped_data_ex2$welfare,
  population = grouped_data_ex2$weight,
  mean = 8,
  pop = NULL,
  p0 = 0.5,
  nobs = 1e5)
str(res)

}
\keyword{internal}
