% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_compute_dist_stats.R
\name{sd_create_synth_vector}
\alias{sd_create_synth_vector}
\title{Create synthetic vector based on Lorenz  parameters of group data}
\usage{
sd_create_synth_vector(
  welfare,
  population,
  mean,
  pop = NULL,
  p0 = 0.5,
  nobs = 100000
)
}
\arguments{
\item{welfare}{numeric: Cumulative proportion of welfare held by that
proportion of the population (Lorenz Curve).}

\item{population}{numeric: Cumulative proportion of population.}

\item{mean}{numeric: Welfare mean.}

\item{p0}{numeric: \strong{TO BE DOCUMENTED}.}

\item{nobs}{numeric: Number of observations to be used in synthetic vector}
}
\value{
list
}
\description{
Compute distributional statistics for grouped data by selecting the best
functional fit for the Lorenz curve (either beta or quadratic).
}
\keyword{internal}
