% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_clean_data.R
\name{standardize_type5}
\alias{standardize_type5}
\title{Standardize type 5}
\usage{
standardize_type5(welfare, population)
}
\arguments{
\item{welfare}{numeric: welfare vector}

\item{population}{numeric: population vector}
}
\value{
data.table
}
\description{
Standardize type 5
}
