% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lb.R
\name{value_at_lb}
\alias{value_at_lb}
\title{Solves for beta Lorenz curves}
\usage{
value_at_lb(x, A, B, C)
}
\arguments{
\item{x}{numeric: point on curve}

\item{A}{numeric vector: First lorenz curve coefficient}

\item{B}{numeric vector: Second lorenz curve coefficient}

\item{C}{numeric vector: Third lorenz curve coefficient}
}
\value{
numeric
}
\description{
\code{value_at_lb()}solves for beta Lorenz curves
}
