% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_compute_pip_stats_lq.R
\name{value_at_lq}
\alias{value_at_lq}
\title{Solves for quadratic Lorenz curves}
\usage{
value_at_lq(x, A, B, C)
}
\arguments{
\item{x}{numeric: Point on curve.}

\item{A}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[1]}.}

\item{B}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[2]}.}

\item{C}{numeric: Lorenz curve coefficient. Output of
\code{regres_lq()$coef[3]}.}
}
\value{
numeric
}
\description{
\code{value_at_lq()}solves for quadratic Lorenz curves with c = 1
General quadratic form: ax^2 + bxy + cy^2 + dx + ey + f = 0
}
\keyword{internal}
