% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gd_estimate_lq.R
\name{value_at_lq}
\alias{value_at_lq}
\title{Solves for quadratic Lorenz curves}
\usage{
value_at_lq(x, A, B, C)
}
\arguments{
\item{x}{numeric: point on curve}

\item{A}{numeric vector: First lorenz curve coefficient}

\item{B}{numeric vector: Second lorenz curve coefficient}

\item{C}{numeric vector: Third lorenz curve coefficient}
}
\value{
numeric
}
\description{
\code{value_at_lq()}solves for quadratic Lorenz curves with c = 1
General quadratic form: ax^2 + bxy + cy^2 + dx + ey + f = 0
}
