% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbpip-package.R
\docType{package}
\name{wbpip-package}
\alias{wbpip}
\alias{wbpip-package}
\title{wbpip: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pip-technical-team.github.io/wbpip}
  \item \url{https://github.com/PIP-Technical-Team/wbpip}
  \item Report bugs at \url{https://github.com/PIP-Technical-Team/wbpip/issues}
}

}
\author{
\strong{Maintainer}: Tony Fujs \email{tfujs@worldbank.org}

Authors:
\itemize{
  \item Aleksander Eilertsen \email{aeilertsen@worldbank.org}
  \item R. Andrés Castañeda \email{acastanedaa@worldbank.org}
  \item Ifeanyi Nzegwu Edochie \email{iedochie@worldbank.org}
}

Other contributors:
\itemize{
  \item Minh Cong Nguyen \email{mnguyen3@worldbank.org} [contributor]
  \item World Bank [copyright holder]
}

}
\keyword{internal}
