% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/low-hanging_indicators.R
\name{which_na}
\alias{which_na}
\alias{which_negative}
\title{find unique obs with specific characteristics}
\usage{
which_na(l)

which_negative(l)
}
\arguments{
\item{l}{list of vectors}
}
\value{
numeric vector with index
}
\description{
\code{which_na} finds out which observations are NA in a list of vectors
\code{which_negative} finds out which observations are negative in a list of vectors
}
\section{Functions}{
\itemize{
\item \code{which_negative()}: finds out which observations are negative in a list of vectors

}}
\keyword{internal}
